% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_deriv_output.R
\name{add_deriv_output}
\alias{add_deriv_output}
\title{Add derived output variables to model output}
\usage{
add_deriv_output(out_list, hyps, vars_sim = NULL)
}
\arguments{
\item{out_list}{list of model output variables}

\item{hyps}{data frame with hypsograph data}

\item{vars_sim}{character vector of variable names to add. If NULL, all variables
with registered functions will be added. If not NULL, only the specified variables
will be added, but their dependencies will also be added.}
}
\value{
List of model output variables with derived variables added
}
\description{
Add derived output variables to model output
}
