% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_humidity_vars.R
\name{add_hum_vars}
\alias{add_hum_vars}
\title{Add humidity variables to meteorological data frame}
\usage{
add_hum_vars(data, hum_method = 1)
}
\arguments{
\item{data}{Data frame with meteorological data including columns:
"hum" (humidity input), "airp" (air pressure, Pa), "sst" (sea surface
temperature, degC or K), "airt" (air temperature, degC or K)}

\item{hum_method}{Method for humidity input:
1 = relative humidity (%)
2 = wet bulb temperature (degC or K)
3 = dew point temperature (degC or K)
4 = specific humidity (kg/kg)}
}
\value{
Data frame with added columns:
\item{es}{Saturation vapour pressure at sea surface temperature (Pa)}
\item{qs}{Saturation specific humidity at sea surface temperature (kg/kg)}
\item{ea}{Actual vapour pressure (Pa)}
\item{qa}{Actual specific humidity (kg/kg)}
\item{rhoa}{Air density (kg/m3)}
}
\description{
Add humidity variables to meteorological data frame
}
