% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_inflows.R
\name{add_inflows}
\alias{add_inflows}
\title{Add inflows to Aeme object}
\usage{
add_inflows(aeme, data)
}
\arguments{
\item{aeme}{Aeme object.}

\item{data}{list with data frames for each inflow. Each data frame must have
columns "Date", "HYD_flow", "HYD_temp" and "CHM_salt". If NULL, no inflows
are added.}
}
\value{
Aeme object with inflows added
}
\description{
Add inflows to Aeme object
}
