% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outflows.R
\name{add_outflows}
\alias{add_outflows}
\title{Add outflows to Aeme object}
\usage{
add_outflows(aeme, data = NULL, elevation = NULL, factor = NULL)
}
\arguments{
\item{aeme}{aeme; object.}

\item{data}{named list with data frames for each outflow. Each data frame 
must have columns "Date", "outflow". If NULL, no outflows are added.}

\item{elevation}{named list with elevation of each outflow. If NULL, no
outflows are added. If elevation is -1, the outflow is assumed to
be at the same elevation as the hypsograph. If elevation is not -1, it must be
within the range of the hypsograph elevation.}

\item{factor}{named list with scaling factors to apply to outflows per model.
If NULL, no scaling is applied.}
}
\value{
Aeme object with outflows added
}
\description{
Add outflows to Aeme object
}
