% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_param.R
\name{add_param}
\alias{add_param}
\title{Add model parameters to Aeme object}
\usage{
add_param(aeme, param)
}
\arguments{
\item{aeme}{Aeme object.}

\item{param}{data frame with columns "model", "file", "name", "value", "min",
"max", "module" and "group.}
}
\value{
Aeme object with parameters added
}
\description{
Add model parameters to Aeme object
}
