% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pars.R
\name{add_pars}
\alias{add_pars}
\title{Add model parameters to Aeme object}
\usage{
add_pars(aeme, pars)
}
\arguments{
\item{aeme}{aeme; object.}

\item{pars}{data frame with columns "model", "file", "name", "value", "min",
"max", "module" and "group.}
}
\value{
Aeme object with parameters added
}
\description{
Add model parameters to Aeme object
}
