% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AEME-package.R
\docType{package}
\name{AEME-package}
\alias{AEME}
\alias{AEME-package}
\title{AEME: Aquatic Ecosystem Model Ensemble}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Run an ensemble of aquatic ecosystem models on lakes and reservoirs using varying levels of inputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/limnotrack/AEME}
  \item \url{https://limnotrack.com/AEME}
}

}
\author{
\strong{Maintainer}: Tadhg Moore \email{tadhg@limnotrack.com} (\href{https://orcid.org/0000-0002-3834-8868}{ORCID})

Authors:
\itemize{
  \item Tadhg Moore \email{tadhg@limnotrack.com} (\href{https://orcid.org/0000-0002-3834-8868}{ORCID})
  \item Christopher McBride \email{chris@limnotrack.com} (\href{https://orcid.org/0000-0001-5784-9767}{ORCID})
}

}
\keyword{internal}
