% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aeme_parameters_bgc}
\alias{aeme_parameters_bgc}
\title{Example dataframe used for calibrating the biogeochemistry in the AEME
models.}
\format{
## `aeme_parameters_bgc`
A data frame with 30 rows and 7 columns:
\describe{
  \item{model}{Model for the parameter}
  \item{file}{File in which the parameter is stored}
  \item{name}{Name of the parameter}
  \item{value}{Value of the parameter}
  \item{min}{Minimum range of the parameter}
  \item{max}{Maximum range of the parameter}
  \item{group}{Phytoplankton Group for the parameter, only applies to
  phytoplankton parameters}
  \item{index}{Index for the parameter in the model file, only used for
  parameters that have multiple values in a vector such as
   "sediment/sed_temp_mean" in GLM-AED2}
  \item{module}{Module for the parameter in the model, useful to help
  identify parameters} 
}
}
\source{
Package development.
}
\usage{
aeme_parameters_bgc
}
\description{
An example dataframe used for inputting and calibrating AEME models.
}
\keyword{datasets}
