% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AemeSummary.R
\docType{class}
\name{AemeSummary}
\alias{AemeSummary}
\alias{AemeSummary-class}
\title{Aeme Summary Class}
\description{
S4 Class representing summarised AEME data
}
\details{
This class represents data summarising an Aeme object.
}
\section{Slots}{

\describe{
\item{\code{aeme}}{An Aeme object which has been summarised}
}}

