% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dycd.R
\name{build_dycd}
\alias{build_dycd}
\title{Build DYRESM-CAEDYM configuration}
\usage{
build_dycd(
  lakename,
  mod_ctrls,
  date_range,
  gps,
  inf,
  outf,
  met,
  hyps,
  lvl,
  lake_dir,
  inf_factor = 1,
  outf_factor = 1,
  ext_elev = 0,
  Kw,
  init_prof,
  use_bgc
)
}
\arguments{
\item{lakename}{string; for lake name}

\item{mod_ctrls}{dataframe; with model controls}

\item{date_range}{vector; of dates}

\item{gps}{vector; of latitude and longitude}

\item{inf}{list of inflows}

\item{outf}{dataframe; of outflow}

\item{met}{dataframe; of meteorological data}

\item{hyps}{dataframe; with hypsography data}

\item{lvl}{dataframe; with lake level data}

\item{lake_dir}{filepath; for outputting model configuration}

\item{inf_factor}{numeric; scaling factor for inflows}

\item{outf_factor}{numeric; scaling factor for outflows}

\item{ext_elev}{numeric; to extend the elevation of the hypsograph}

\item{Kw}{numeric; light extinction coefficient}

\item{init_prof}{dataframe; of initial profile with depth, temperature and
salinity.}

\item{use_bgc}{boolean; switch for biogrochemistry model.}
}
\description{
Build DYRESM-CAEDYM configuration
}
