% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_glm.R
\name{build_glm}
\alias{build_glm}
\title{Build a glm-aed model from generic inputs}
\usage{
build_glm(
  lakename,
  mod_ctrls,
  date_range,
  lake_shape,
  gps,
  hyps,
  lvl,
  inf,
  outf,
  met,
  lake_dir,
  config_dir,
  init_prof,
  inf_factor = 1,
  outf_factor = 1,
  Kw,
  ext_elev,
  key_naming,
  use_bgc,
  use_lw
)
}
\arguments{
\item{lake_shape}{shapefile}

\item{use_lw}{}
}
\description{
Build a glm-aed model from generic inputs
}
