% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lake_vol.R
\name{calc_lake_vol}
\alias{calc_lake_vol}
\title{Calculate the volume of a lake using bathymetry data or a hypsograph}
\usage{
calc_lake_vol(aeme = NULL, hyps = NULL, depth = 0)
}
\arguments{
\item{aeme}{Aeme object.}

\item{hyps}{data.frame with columns 'depth' and 'area'.}

\item{depth}{numeric. The depth to which to calculate the volume. If
provided, the volume will be calculated to this depth. If not provided, the
volume will be calculated to the maximum depth of the hypsograph.}
}
\value{
numeric. The volume of the lake in cubic meters (m^3).
}
\description{
Calculate the volume of a lake using bathymetry data or a hypsograph
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
vol <- calc_lake_vol(aeme)
}
