% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tli_helpers.R
\name{calc_tli3}
\alias{calc_tli3}
\title{Calculate TLI 3}
\usage{
calc_tli3(chla, tn, tp)
}
\arguments{
\item{chla}{Chlorophyll-a concentration in µg/L}

\item{tn}{Total nitrogen concentration in mg/L}

\item{tp}{Total phosphorus concentration in mg/L}
}
\value{
TLI 3
}
\description{
Calculate TLI 3
}
