% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tli_helpers.R
\name{calc_tli4}
\alias{calc_tli4}
\title{Calculate TLI 4}
\usage{
calc_tli4(chla, tn, tp, secchi)
}
\arguments{
\item{chla}{Chlorophyll-a concentration in µg/L}

\item{tn}{Total nitrogen concentration in mg/L}

\item{tp}{Total phosphorus concentration in mg/L}

\item{secchi}{Secchi depth in meters}
}
\value{
TLI 4
}
\description{
Calculate TLI 4
}
