% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tli_helpers.R
\name{calc_tli_chla}
\alias{calc_tli_chla}
\title{Calculate TLI chlorophyll component}
\usage{
calc_tli_chla(chla)
}
\arguments{
\item{chla}{Chlorophyll-a concentration in µg/L}
}
\value{
TLI chlorophyll component
}
\description{
Calculate TLI chlorophyll component
}
