% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tli_helpers.R
\name{calc_tli_n}
\alias{calc_tli_n}
\title{Calculate TLI nitrogen component}
\usage{
calc_tli_n(tn)
}
\arguments{
\item{tn}{Total nitrogen concentration in mg/L}
}
\value{
TLI nitrogen component
}
\description{
Calculate TLI nitrogen component
}
