% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tli_helpers.R
\name{calc_tli_p}
\alias{calc_tli_p}
\title{Calculate TLI phosphorus component}
\usage{
calc_tli_p(tp)
}
\arguments{
\item{tp}{Total phosphorus concentration in mg/L}
}
\value{
TLI phosphorus component
}
\description{
Calculate TLI phosphorus component
}
