% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tli_helpers.R
\name{calc_tli_secchi}
\alias{calc_tli_secchi}
\title{Calculate TLI secchi component}
\usage{
calc_tli_secchi(secchi)
}
\arguments{
\item{secchi}{Secchi depth in meters}
}
\value{
TLI secchi component
}
\description{
Calculate TLI secchi component
}
