% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wbal.R
\name{calc_wbal}
\alias{calc_wbal}
\title{Estimate lake water balance based on a minimal set of inputs}
\usage{
calc_wbal(
  aeme_time,
  model,
  use,
  hyps,
  inf,
  outf = NULL,
  level = NULL,
  obs_lake = NULL,
  obs_met,
  ext_elev,
  elevation,
  print_plots = FALSE,
  coeffs = NULL
)
}
\arguments{
\item{aeme_time}{list; time object from aeme_object using `time()`}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{hyps}{data frame of hypsographic curve, elevation (masl) and planar
area (m^2)}

\item{inf}{list of inflow data frames}

\item{outf}{list of outflow data frames. Default = NULL}

\item{level}{data frame of lake water level observations.. cols = Date,
value}

\item{obs_lake}{data frame of lake observations in ensemble standard format}

\item{obs_met}{data frame of meteorology, must include MET_tmpair, MET_wndspd
& MET_prvapr, continuous Date, extent defines output extent}

\item{ext_elev}{numeric; elevation to extend bathymetry}

\item{elevation}{numeric; elevation of lake}

\item{print_plots}{logical; print plots of water balance components}

\item{coeffs}{numeric vector; coefficients for estimating lake surface
temperature. Default = NULL}

\item{method}{character; method for estimating lake water level. Options
are "use_obs", "sinusoidal", "constant".}
}
\value{
data frame of water balance components which are:
- Date
- lvlwtr
- HYD_flow
-...
}
\description{
Estimate lake water balance based on a minimal set of inputs
}
