% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{catchment,aeme-method}
\alias{catchment,aeme-method}
\title{Access catchment slot}
\usage{
\S4method{catchment}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of catchment characteristics
}
\description{
Access catchment slot
}
