% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{catchment<-,aeme-method}
\alias{catchment<-,aeme-method}
\title{Set catchment in aeme object}
\usage{
\S4method{catchment}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New catchment data to be assigned.}
}
\value{
A modified aeme object with updated catchment slot.
}
\description{
Update the catchment slot of an aeme object
}
\details{
This method updates the "catchment" slot of a aeme object with new data.
}
