% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{catchment<-}
\alias{catchment<-}
\title{Set catchment in aeme object}
\usage{
catchment(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New catchment data to be assigned.}
}
\value{
A modified aeme object with updated catchment slot.
}
\description{
Update the catchment slot of an aeme object
}
