% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{catchment}
\alias{catchment}
\title{Access catchment slot}
\usage{
catchment(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of catchment characteristics
}
\description{
Access catchment slot
}
