% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_AEME_pkg.R
\name{check_AEME_pkg}
\alias{check_AEME_pkg}
\title{Check if the package is working correctly}
\usage{
check_AEME_pkg()
}
\value{
TRUE if the package is working correctly
}
\description{
Check if the package is working correctly
}
