% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_aeme_vars.R
\name{check_aeme_vars}
\alias{check_aeme_vars}
\title{Check AEME variable names}
\usage{
check_aeme_vars(x)
}
\arguments{
\item{x}{Character vector of variable names to check.}
}
\value{
Invisibly returns TRUE if all variables are valid, otherwise throws 
an error.
}
\description{
Check if the provided variable names are valid AEME variable names.
}
\examples{
check_aeme_vars("HYD_temp")
}
