% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_glm_nml.R
\name{check_glm_nml}
\alias{check_glm_nml}
\title{Check GLM nml for common issues}
\usage{
check_glm_nml(file)
}
\arguments{
\item{file}{path to GLM nml file}
}
\value{
Invisibly returns TRUE if no issues found, otherwise aborts with informative messages
}
\description{
Check GLM nml for common issues
}
