% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check model name and return standardized code}
\usage{
check_model(model, os_valid = FALSE)
}
\arguments{
\item{model}{Character vector of model names. Valid options are:
"DYRESM-CAEDYM", "GLM-AED", "GOTM-WET", "SIMSTRAT-AED2" or their
corresponding codes "dy_cd", "glm_aed", "gotm_wet", "simstrat_aed2".}

\item{os_valid}{Logical. If TRUE, checks if the model is valid for the
current operating system.}
}
\value{
Character vector of standardized model codes.
}
\description{
Check model name and return standardized code
}
\examples{
check_model(c("GLM-AED", "gotm_wet"))
}
