% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_obs_var.R
\name{check_obs_var}
\alias{check_obs_var}
\title{Check if a variable is present in the observations for the model time period}
\usage{
check_obs_var(aeme, var_sim)
}
\arguments{
\item{aeme}{Aeme object.}

\item{var_sim}{character; variable in the AEME format (e.g. "HYD_temp").}
}
\value{
A list with two elements:
\itemize{
\item \code{obs}: A data frame with the following columns:
\itemize{
\item \code{model}: Model name
\item \code{var_aeme}: Variable name
\item \code{n}: Number of observations
}
\item \code{vars_present}: A character vector of variables present in the
observations
}
}
\description{
Check if a variable is present in the observations for the model time period
}
