% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path.R
\name{check_path}
\alias{check_path}
\title{Check and manage file paths}
\usage{
check_path(path, create = FALSE, must_exist = FALSE)
}
\arguments{
\item{path}{Character string specifying the file path to check.}

\item{create}{Logical indicating whether to create the directory if it does
not exist. Default is FALSE.}

\item{must_exist}{Logical indicating whether the directory must exist. If
TRUE and the directory does not exist, an error is thrown. Default is FALSE.}
}
\value{
Normalized file path as a character string.
}
\description{
Check and manage file paths
}
\examples{
check_path("aeme", create = TRUE)
}
