% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{cli_inform_safe}
\alias{cli_inform_safe}
\title{Inform messages respecting the global AEME.inform option}
\usage{
cli_inform_safe(...)
}
\arguments{
\item{...}{arguments passed to cli_inform_safe()}
}
\description{
Inform messages respecting the global AEME.inform option
}
