% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{cli_safe}
\alias{cli_safe}
\title{Inform messages respecting the global AEME.inform option}
\usage{
cli_safe(..., FUN = cli::cli_bullets, indent = TRUE)
}
\arguments{
\item{...}{arguments passed to cli_inform_safe()}

\item{FUN}{function to use for messaging, default is cli::cli_inform}

\item{indent}{logical, whether to indent the message, default is FALSE}
}
\description{
Used primarily as an internal helper to safely suppress messages to console.
Messages are printed if the global option is set to TRUE:
\code{options(AEME.inform = TRUE)}
}
