% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{configuration,aeme-method}
\alias{configuration,aeme-method}
\title{Access configuration slot}
\usage{
\S4method{configuration}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of configuration characteristics
}
\description{
Access configuration slot
}
