% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{configuration<-,Aeme-method}
\alias{configuration<-,Aeme-method}
\title{Set configuration in Aeme object}
\usage{
\S4method{configuration}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New configuration data to be assigned.}
}
\value{
A modified Aeme object with updated configuration slot.
}
\description{
Update the configuration slot of an aeme object
}
\details{
This method updates the "configuration" slot of An Aeme object with new data.
}
