% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{configuration}
\alias{configuration}
\alias{configuration,Aeme-method}
\title{Access the configuration slot of an Aeme object}
\usage{
configuration(aeme)

\S4method{configuration}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
List of configuration characteristics.
}
\description{
Access the configuration slot of an Aeme object
}
