% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_do.R
\name{convert_do}
\alias{convert_do}
\title{Convert dissolved oxygen between mg/L and percent saturation}
\usage{
convert_do(
  value,
  depth,
  temp,
  baro,
  altitude = 0,
  salinity = 0,
  model = "garcia_benson",
  direction = "to_mgL"
)
}
\arguments{
\item{value}{numeric vector of dissolved oxygen values to convert}

\item{depth}{depth (m) at which the DO measurement was made}

\item{temp}{water temperature (°C)}

\item{baro}{barometric pressure (mb)}

\item{altitude}{altitude (m). Only used if `baro` is missing.}

\item{salinity}{salinity (ppt). Default is 0 for freshwater.}

\item{model}{character, solubility model to use. Options are
"garcia", "garcia_benson", "weiss", or "benson".}

\item{direction}{character, conversion direction. Options are
"to_mgL" (percent saturation to mg/L) or "to_percent" (mg/L to percent
saturation).}
}
\value{
numeric vector of converted dissolved oxygen values
}
\description{
Convert dissolved oxygen between mg/L and percent saturation
}
