% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_era5.R
\name{download_era5}
\alias{download_era5}
\title{Download ERA5}
\usage{
download_era5(
  lat,
  lon,
  variable = c("10m_u_component_of_wind", "10m_v_component_of_wind",
    "2m_dewpoint_temperature", "2m_temperature", "snowfall", "surface_pressure",
    "surface_solar_radiation_downwards", "surface_thermal_radiation_downwards",
    "total_precipitation"),
  year = 2022,
  month = 1:12,
  site = "test",
  user = NULL,
  era5_dataset = "reanalysis-era5-land",
  path = ".",
  job = TRUE
)
}
\arguments{
\item{lat}{numeric; latitude}

\item{lon}{numeric; longitude}

\item{variable}{string with ERA5 variable names e.g. "2m_temperature", "total_precipitation"}

\item{year}{numeric; year or vector of years.}

\item{month}{numeric; month or vector of months. Defaults to 1:12.}

\item{site}{string of site name which will be appended to the file}

\item{user}{user ID linked with Copernicus account}

\item{era5_dataset}{string of which ERA5 dataset to use. Can be 'reanalysis-era5-single-levels' or 'reanalysis-era5-land'}

\item{path}{filepath to store downloaded file}

\item{job}{logical; send the request to a background job in RStudio. Only
works in RStudio. Default = TRUE.}
}
\description{
Download ERA5 meteorological data from the Copernicus Data Store (CDS).
Create a ECMWF account by self registering (https://accounts.ecmwf.int/auth/realms/ecmwf/login-actions/registration?client_id=cms-www&tab_id=kZuwr5qOsFM) and retrieving your key at https://api.ecmwf.int/v1/key/ after
you log in.
}
