% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_met.R
\name{expand_met}
\alias{expand_met}
\title{Expand a minimal set of meteorology inputs to a complete set of variables suitable for all models}
\usage{
expand_met(met, coords.xyz, print.plot = FALSE)
}
\arguments{
\item{met}{data frame of source meteorology variables}

\item{coords.xyz}{vector of longitude, latitude and elevation of met station}

\item{print.plot}{boolean; print the plot}
}
\value{
dataframe with expanded met variables
}
\description{
Expand a minimal set of meteorology inputs to a complete set of variables suitable for all models
}
