% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_value.R
\name{get_config_value}
\alias{get_config_value}
\title{Get a configuration value from the aeme object. If the key is not present,
return the default value.}
\usage{
get_config_value(aeme, key, default = NULL)
}
\arguments{
\item{aeme}{Aeme object.}

\item{key}{The name of the configuration value to retrieve.}

\item{default}{The default value to return if the key is not present in the
configuration. If NULL, the default value from config_defaults() will be used.
If the key is not present in config_defaults(), NULL will be returned.}
}
\value{
The value of the configuration key, or the default value if the key
is not present in the configuration.
}
\description{
Get a configuration value from the aeme object. If the key is not present,
return the default value.
}
