% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deriv_vars.R
\name{get_deriv_inputs}
\alias{get_deriv_inputs}
\title{Get derived variables needed for simulation}
\usage{
get_deriv_inputs(vars_sim)
}
\arguments{
\item{vars_sim}{Character vector of variable names being simulated}
}
\value{
Character vector of variable names that are needed as inputs
}
\description{
Get derived variables needed for simulation
}
\examples{
get_deriv_inputs(vars_sim  = c("HYD_thmcln", "HYD_epidep", "CHM_oxyepi", "LKE_tli4"))
}
