% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glm_sed_params.R
\name{get_glm_sed_params}
\alias{get_glm_sed_params}
\title{Get number of sediment zones in GLM-AED model}
\usage{
get_glm_sed_params(aeme, path, lake_dir = NULL)
}
\arguments{
\item{aeme}{aeme; object.}

\item{path}{filepath; where input files are located relative to the current
working directory.}

\item{lake_dir}{Path to the lake AEME directory. If NULL, it will be
computed from \code{aeme} and \code{path}.}
}
\value{
Number of sediment zones
}
\description{
Get number of sediment zones in GLM-AED model
}
