% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hm_palette.R
\name{get_hm_palette}
\alias{get_hm_palette}
\title{Get default heatmap palette for a variable}
\usage{
get_hm_palette(var, n = NULL)
}
\arguments{
\item{var}{Character. Variable name.}

\item{n}{Number of colours required.}
}
\value{
A character vector of colours
}
\description{
Get default heatmap palette for a variable
}
