% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hypsograph.R
\name{get_hypsograph}
\alias{get_hypsograph}
\title{Get hypsograph from Aeme object}
\usage{
get_hypsograph(aeme)
}
\arguments{
\item{aeme}{Aeme object.}
}
\value{
Hypsograph data frame or NULL if not present
}
\description{
Get hypsograph from Aeme object
}
