% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake.R
\name{get_lake}
\alias{get_lake}
\title{Get lake data from Aeme object}
\usage{
get_lake(aeme)
}
\arguments{
\item{aeme}{Aeme object.}
}
\value{
Lake data frame or NULL if not present.
}
\description{
Get lake data from Aeme object
}
