% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake_dir.R
\name{get_lake_dir}
\alias{get_lake_dir}
\title{Get the directory of the lake model setup}
\usage{
get_lake_dir(aeme, path)
}
\arguments{
\item{aeme}{Aeme object.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}
}
\value{
character; the directory of the lake model setup
}
\description{
Get the directory of the lake model setup
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
path <- tempdir()
model <- c("glm_aed")
aeme <- build_aeme(aeme = aeme, model = model, path = path, ext_elev = 3)
lake_dir <- get_lake_dir(aeme)

}
