% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met.R
\name{get_met}
\alias{get_met}
\title{Get meteorological data from Aeme object}
\usage{
get_met(aeme)
}
\arguments{
\item{aeme}{Aeme object.}
}
\value{
Meteorological data frame or NULL if not present.
}
\description{
Get meteorological data from Aeme object
}
