% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_vars.R
\name{get_met_vars}
\alias{get_met_vars}
\title{Get the names and units for the meteorological variables}
\usage{
get_met_vars(as_vector = FALSE)
}
\arguments{
\item{as_vector}{Logical. If TRUE, returns a named vector with column names
for the meteorological variables. If FALSE, returns a data frame with v
ariable names, column names and units. Default is FALSE.}
}
\value{
Named vector with column names for the meteorological variables.
The names of the vector are the variable names in the Aeme object and the
values are the corresponding column names in the input data frame.
}
\description{
Get the names and units for the meteorological variables
}
\examples{
get_met_vars()

# Get the column names for the meteorological variables as a named vector
met_cols <- get_met_vars(as_vector = TRUE)

}
