% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_controls.R
\name{get_model_controls}
\alias{get_model_controls}
\title{Get model controls}
\usage{
get_model_controls(aeme = NULL, use_bgc = FALSE)
}
\arguments{
\item{aeme}{Aeme object.}

\item{use_bgc}{logical; enable the biogeochemical model. Default: \code{FALSE}.}
}
\value{
A data frame of model controls
}
\description{
Get model controls
}
\examples{
\dontrun{
model_controls <- get_model_controls()
}

}
