% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_dir.R
\name{get_model_dir}
\alias{get_model_dir}
\title{Get the model directory for a given AEME and model.}
\usage{
get_model_dir(aeme, model, path)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}, \code{"simstrat_aed2"}. Defaults to all models if
not found in \code{aeme}.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}
}
\value{
Path to the model directory
}
\description{
Get the model directory for a given AEME and model.
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
path <- tempdir()
model <- c("glm_aed")
aeme <- build_aeme(aeme = aeme, model = model, path = path, ext_elev = 3)
lake_dir <- get_lake_dir(aeme)
model_dir <- get_model_dir(aeme)

}
