% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_models.R
\name{get_models}
\alias{get_models}
\title{Get a vector of available models}
\usage{
get_models(aeme = NULL)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
vector of available models
}
\description{
Get a vector of available models
}
