% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Get AEME output and add to the aeme object}
\usage{
get_output(model, aeme_data, path, mod_ctrls, parallel = FALSE, nlev = NULL)
}
\arguments{
\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{aeme_data}{aeme; data object.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{mod_ctrls}{dataframe; of configuration loaded from
"mod_ctrls.csv".}

\item{parallel}{boolean; run models in parallel. Defaults to FALSE.}

\item{nlev}{}
}
\value{
Updated aeme object with model output
}
\description{
Get AEME output and add to the aeme object
}
