% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_vars.R
\name{get_output_vars}
\alias{get_output_vars}
\title{Get the output variables from an AEME object}
\usage{
get_output_vars(
  aeme,
  model,
  ens_n = 1,
  group = c("core", "diag", "sediment")[1]
)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}, \code{"simstrat_aed2"}. Defaults to all models if
not found in \code{aeme}.}

\item{ens_n}{numeric; ensemble number to allocate to model output which is
loaded. Defaults to 1.}

\item{group}{character; which variables to list: \code{"core"} (time and
depth x time variables, default), \code{"diag"} (diagnostics and scalars),
and/or \code{"sediment"} (sediment zone/layer variables). Diagnostic and
sediment variables keep their raw model names.}
}
\value{
A character vector of the output variables
}
\description{
Get the output variables from an AEME object
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
path <- tempdir()
model_controls <- get_model_controls(use_bgc = TRUE)
model <- c("glm_aed")
aeme <- build_aeme(path = path, aeme = aeme, model = model,
                   model_controls = model_controls,
                   ext_elev = 5, use_bgc = TRUE)
# Run models
aeme <- run_aeme(aeme = aeme, model = model, verbose = FALSE,
path = path, model_controls = model_controls)
get_output_vars(aeme, model)
}
