% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_station_pressure}
\alias{get_station_pressure}
\title{Return station pressure from mean sea level pressure.}
\usage{
get_station_pressure(prmslp, elevation, tmpair)
}
\arguments{
\item{prmslp}{A numeric vector of mean sea level pressure in Pa}

\item{elevation}{A numeric vector of elevation in m}

\item{tmpair}{A numeric vector of air temperature in degC}
}
\value{
A numeric vector of station pressure in Pa
}
\description{
Return station pressure from mean sea level pressure.
}
\note{
This function is just the inverse of \code{\link{get_mean_sea_level_pressure}}.
}
\examples{
get_station_pressure(101226.5, 105:205, 17.19)

}
\references{
See \code{\link{get_mean_sea_level_pressure}}.
}
