% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_to_aeme.R
\name{glm_to_aeme}
\alias{glm_to_aeme}
\title{Load a GLM nml file and convert to aeme object}
\usage{
glm_to_aeme(nml_file)
}
\arguments{
\item{nml_file}{Path to GLM nml file}
}
\value{
Aeme object
}
\description{
Load a GLM nml file and convert to aeme object
}
