% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{inflows,Aeme-method}
\alias{inflows,Aeme-method}
\title{Access inflows slot}
\usage{
\S4method{inflows}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
list of inflows characteristics
}
\description{
Access inflows slot
}
