% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{inflows<-,aeme-method}
\alias{inflows<-,aeme-method}
\title{Set inflows in aeme object}
\usage{
\S4method{inflows}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New inflows data to be assigned.}
}
\value{
A modified aeme object with updated inflows slot.
}
\description{
Update the inflows slot of an aeme object
}
\details{
This method updates the "inflows" slot of a aeme object with new data.
}
