% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{inflows<-}
\alias{inflows<-}
\alias{inflows<-,Aeme-method}
\title{Set the inflows slot of an Aeme object}
\usage{
inflows(aeme) <- value

\S4method{inflows}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New inflows list to assign.}
}
\value{
A modified Aeme object with updated inflows slot.
}
\description{
Set the inflows slot of an Aeme object
}
