% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{input,Aeme-method}
\alias{input,Aeme-method}
\title{Access input slot}
\usage{
\S4method{input}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
list of input characteristics
}
\description{
Access input slot
}
