% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{input<-,aeme-method}
\alias{input<-,aeme-method}
\title{Set input in aeme object}
\usage{
\S4method{input}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New input data to be assigned.}
}
\value{
A modified aeme object with updated input slot.
}
\description{
Update the input slot of an aeme object
}
\details{
This method updates the "input" slot of a aeme object with new data.
}
