% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{input}
\alias{input}
\alias{input,Aeme-method}
\title{Access the input slot of an Aeme object}
\usage{
input(aeme)

\S4method{input}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
List of input characteristics.
}
\description{
Access the input slot of an Aeme object
}
