% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_aeme.R
\name{insert_aeme}
\alias{insert_aeme}
\title{Insert default aeme list code snippet.}
\usage{
insert_aeme()
}
\value{
Inserts code snippet at cursor position
}
\description{
Call this function as an addin to insert a default \code{aeme} list at the
cursor position.
}
