% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme_output.R
\name{is_aeme_output}
\alias{is_aeme_output}
\title{Is \code{x} a model-output list (standardised or raw)?}
\usage{
is_aeme_output(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
logical.
}
\description{
Is \code{x} a model-output list (standardised or raw)?
}
