% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme_output.R
\name{is_aeme_output_raw}
\alias{is_aeme_output_raw}
\title{Is \code{x} a \emph{raw} (\code{raw_output = TRUE}) model-output list?}
\usage{
is_aeme_output_raw(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
logical.
}
\description{
Is \code{x} a \emph{raw} (\code{raw_output = TRUE}) model-output list?
}
