% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_outputs.R
\name{is_model_error}
\alias{is_model_error}
\title{Check if object is a model output error}
\usage{
is_model_error(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if object is a model output error
}
\description{
Check if object is a model output error
}
