% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{lake<-,aeme-method}
\alias{lake<-,aeme-method}
\title{Set lake in aeme object}
\usage{
\S4method{lake}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New lake data to be assigned.}
}
\value{
A modified aeme object with updated lake slot.
}
\description{
Update the lake slot of an aeme object
}
\details{
This method updates the "lake" slot of a aeme object with new data.
}
