% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{lake}
\alias{lake}
\alias{lake,Aeme-method}
\title{Access the lake slot of an Aeme object}
\usage{
lake(aeme)

\S4method{lake}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
List of lake characteristics.
}
\description{
Access the lake slot of an Aeme object
}
