% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{Get a vector of available models}
\usage{
list_models(aeme = NULL)
}
\arguments{
\item{aeme}{Aeme object.}
}
\value{
vector of available models
}
\description{
Get a vector of available models
}
