% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DYCDcfg.R
\name{make_DYCDcfg}
\alias{make_DYCDcfg}
\title{Make a cfg file for dyresm(caedym) simulation}
\usage{
make_DYCDcfg(
  lakename = "unknown",
  date_range,
  verDY = 3.1,
  runCD = FALSE,
  EXTC = 0.2,
  minLyrThk = 0.2,
  maxLyrThk = 0.3,
  simVars,
  filePath
)
}
\arguments{
\item{lakename}{name of lake being simulated}

\item{date_range}{vector of start and end dates e.g., c(YYYY-MM-DD, YYYY-MM-DD)}

\item{verDY}{dyresm version number}

\item{runCD}{turn CAEDYM on (TRUE) or off (FALSE)}

\item{EXTC}{light extinction coefficient (only used if runCD == FALSE)}

\item{minLyrThk}{minimum layer thickness (default 0.2)}

\item{maxLyrThk}{maximum layer thickness (default 3.0)}

\item{simVars}{variables to be simulated (default = DYRESM vars only)}

\item{filePath}{dir of model, relative to working dir}
}
\description{
Make a cfg file for dyresm(caedym) simulation
}
\keyword{inputs}
