% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DYmet.R
\name{make_DYmet}
\alias{make_DYmet}
\title{Make a .met (meteorology) file for DY(CD) simulation}
\usage{
make_DYmet(
  lakename = "unknown",
  info = "",
  verDY = 3.1,
  obsMet,
  infRain = FALSE,
  wndType = 0,
  metHeight = 10,
  z_max,
  filePath = ""
)
}
\arguments{
\item{lakename}{name of lake being simulated}

\item{info}{extra text to be printed in file header (base = 'DYRESM meterological inputs file for lake <lake>')}

\item{verDY}{DYRESM version}

\item{obsMet}{data frame of meteorological data in DY order with standard dates (YYYY-MM-DD)}

\item{infRain}{TRUE = set .met rain to zero (to be put in as inflow instead)}

\item{wndType}{0 = fixed height, 1 = floating}

\item{metHeight}{height of met station above lake surface}

\item{filePath}{path to write the complete .met file}

\item{Z_max}{long-term average max depth of lake}
}
\description{
Make a .met (meteorology) file for DY(CD) simulation
}
\keyword{inputs}
