% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_infGLM.R
\name{make_infGLM}
\alias{make_infGLM}
\title{write and configure inflows for a GLM simulation}
\usage{
make_infGLM(glm_nml, path_glm, list_inf, mass = TRUE, inf_factor = 1)
}
\arguments{
\item{inf_factor}{}
}
\description{
write and configure inflows for a GLM simulation
}
