% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stgGLM.R
\name{make_stgGLM}
\alias{make_stgGLM}
\title{Write a lake bathymetry to a GLM simulation}
\usage{
make_stgGLM(glm_nml, lakename, bathy, gps, dims_lake, update_sediment = TRUE)
}
\arguments{
\item{update_sediment}{}
}
\description{
Write a lake bathymetry to a GLM simulation
}
