% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_layer_structure}
\alias{model_layer_structure}
\title{Reference data frame for model layer structure.}
\format{
`model_layer_structure`
A data frame with 191 rows and 3 columns:
\describe{
  \item{zi}{Interface depth (m)}
  \item{h}{Layer thickness (m)}
  \item{z}{Layer depth (m)}
  \item{n}{Layer number}
}
}
\source{
Package development.
}
\usage{
model_layer_structure
}
\description{
The reference used to structure model layers within the different models.
For the gotm_wet model, this is used to estimate the fractions
at different depths. Whereas for the glm_aed and dy_cd models, this is used
to define the min and max width of the layers.
}
\keyword{datasets}
