% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{observations,aeme-method}
\alias{observations,aeme-method}
\title{Access observations slot}
\usage{
\S4method{observations}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of observations characteristics
}
\description{
Access observations slot
}
