% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{observations<-,aeme-method}
\alias{observations<-,aeme-method}
\title{Set observations in aeme object}
\usage{
\S4method{observations}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New observations data to be assigned.}
}
\value{
A modified aeme object with updated observations slot.
}
\description{
Update the observations slot of an aeme object
}
\details{
This method updates the "observations" slot of a aeme object with new data.
}
