% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{observations<-}
\alias{observations<-}
\alias{observations<-,Aeme-method}
\title{Set the observations slot of an Aeme object}
\usage{
observations(aeme) <- value

\S4method{observations}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New observations list to assign.}
}
\value{
A modified Aeme object with updated observations slot.
}
\description{
Set the observations slot of an Aeme object
}
