% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{observations}
\alias{observations}
\alias{observations,Aeme-method}
\title{Access the observations slot of an Aeme object}
\usage{
observations(aeme)

\S4method{observations}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
List of observations characteristics.
}
\description{
Access the observations slot of an Aeme object
}
