% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_output.R
\name{open_nc_safe}
\alias{open_nc_safe}
\title{Safely open a NetCDF file}
\usage{
open_nc_safe(file, model)
}
\arguments{
\item{file}{path to NetCDF file}

\item{model}{name of the model (for error messages)}
}
\value{
ncdf4 object
}
\description{
Safely open a NetCDF file
}
