% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{outflows,aeme-method}
\alias{outflows,aeme-method}
\title{Access outflows slot}
\usage{
\S4method{outflows}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of outflows characteristics
}
\description{
Access outflows slot
}
