% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{outflows<-,aeme-method}
\alias{outflows<-,aeme-method}
\title{Set outflows in aeme object}
\usage{
\S4method{outflows}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New outflows data to be assigned.}
}
\value{
A modified aeme object with updated outflows slot.
}
\description{
Update the outflows slot of an aeme object
}
\details{
This method updates the "outflows" slot of a aeme object with new data.
}
