% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{outflows<-}
\alias{outflows<-}
\alias{outflows<-,Aeme-method}
\title{Set the outflows slot of an Aeme object}
\usage{
outflows(aeme) <- value

\S4method{outflows}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New outflows list to assign.}
}
\value{
A modified Aeme object with updated outflows slot.
}
\description{
Set the outflows slot of an Aeme object
}
