% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{output,Aeme-method}
\alias{output,Aeme-method}
\title{Access output slot}
\usage{
\S4method{output}{Aeme}(aeme)
}
\arguments{
\item{aeme}{A Aeme object.}
}
\value{
list of output characteristics
}
\description{
Access output slot
}
