% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{output<-,Aeme-method}
\alias{output<-,Aeme-method}
\title{Set output in Aeme object}
\usage{
\S4method{output}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New output data to be assigned.}
}
\value{
A modified Aeme object with updated output slot.
}
\description{
Update the output slot of an Aeme object
}
\details{
This method updates the "output" slot of an Aeme object with new data.
}
