% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{output<-}
\alias{output<-}
\alias{output<-,Aeme-method}
\title{Set the output slot of an Aeme object}
\usage{
output(aeme) <- value

\S4method{output}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New output list to assign.}
}
\value{
A modified Aeme object with updated output slot.
}
\description{
Set the output slot of an Aeme object
}
