% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{output}
\alias{output}
\alias{output,Aeme-method}
\title{Access the output slot of an Aeme object}
\usage{
output(aeme)

\S4method{output}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
List of output characteristics.
}
\description{
Access the output slot of an Aeme object
}
