% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_colnames.R
\name{param_colnames}
\alias{param_colnames}
\title{Get AEME parameters data frame column names}
\usage{
param_colnames(incl_opt = TRUE)
}
\arguments{
\item{incl_opt}{Logical; if TRUE, include optional parameter names. Default
is TRUE.}
}
\value{
A character vector of AEME parameter data frame column names.
}
\description{
Get AEME parameters data frame column names
}
