% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{parameters<-,aeme-method}
\alias{parameters<-,aeme-method}
\title{Set parameters in aeme object}
\usage{
\S4method{parameters}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New parameters data to be assigned.}
}
\value{
A modified aeme object with updated parameters slot.
}
\description{
Update the parameters slot of an aeme object
}
\details{
This method updates the "parameters" slot of a aeme object with new data.
}
