% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{parameters}
\alias{parameters}
\alias{parameters,Aeme-method}
\title{Access the parameters slot of an Aeme object}
\usage{
parameters(aeme)

\S4method{parameters}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
Dataframe of parameters.
}
\description{
Access the parameters slot of an Aeme object
}
