% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_check.R
\name{pkg_check}
\alias{pkg_check}
\title{Check if the package is working correctly}
\usage{
pkg_check()
}
\value{
TRUE if the package is working correctly
}
\description{
Check if the package is working correctly
}
