% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{plot,Aeme-method}
\alias{plot,Aeme-method}
\title{Plot method for Aeme objects}
\usage{
\S4method{plot}{Aeme}(x, y, ..., add = FALSE)
}
\arguments{
\item{x}{An Aeme object.}

\item{y}{An Aeme slot name (character). Defaults to \code{"output"}.}

\item{...}{Additional arguments (currently unused).}

\item{add}{Logical; add to current plot? (currently unused)}
}
\value{
A ggplot object, or prints to the active graphics device.
}
\description{
Plot an Aeme object
}
\details{
This method plots the Aeme object.
}
