% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_kw_regime.R
\name{plot.kw_regime}
\alias{plot.kw_regime}
\title{Plot a Kw regime suggestion}
\usage{
\method{plot}{kw_regime}(x, ...)
}
\arguments{
\item{x}{an object of class \code{kw_regime} (see
\code{\link[=suggest_kw_regime]{suggest_kw_regime()}}).}

\item{...}{unused.}
}
\value{
A \code{patchwork} object combining the observed Kw time series and
the monthly climatology.
}
\description{
Plot a Kw regime suggestion
}
