% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_est_wbal.R
\name{plot_est_wbal}
\alias{plot_est_wbal}
\title{Plot estimated water balance diagnostics from an aeme object.}
\usage{
plot_est_wbal(aeme, model, time_axis = c("auto", "daily", "monthly", "annual"))
}
\arguments{
\item{aeme}{aeme object.}

\item{model}{character; single model name. If missing, uses list_models().}

\item{time_axis}{one of \code{"auto"} (default), \code{"daily"},
\code{"monthly"}, or \code{"annual"}.}
}
\value{
A patchwork object, or NULL if no water balance data available.
}
\description{
Plot estimated water balance diagnostics from an aeme object.
}
