% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flows.R
\name{plot_flows}
\alias{plot_flows}
\title{Plot inflows and/or outflows}
\usage{
plot_flows(aeme, flow = c("inflow", "outflow"), var_sim = "HYD_flow")
}
\arguments{
\item{aeme}{Aeme object.}

\item{flow}{character vector; "inflow", "outflow" or both. Default is both.
For outflow, var_sim must be "HYD_flow".}

\item{var_sim}{character; column name in the data frame. Default is
"HYD_flow".}
}
\value{
A ggplot object
}
\description{
Plot inflows and/or outflows
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
plot_flows(aeme = aeme)
}
