% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hyps.R
\name{plot_hyps}
\alias{plot_hyps}
\title{Plot hypsograph}
\usage{
plot_hyps(aeme, y = "elev", add_surface = FALSE, incl_ext_elev = FALSE)
}
\arguments{
\item{aeme}{Aeme object.}

\item{y}{Character; column name in hypsograph data to plot on y-axis. Can
be either 'elev' or 'depth'.}

\item{add_surface}{Logical; if TRUE, add a horizontal line representing the
lake surface elevation. Default is FALSE.}

\item{incl_ext_elev}{Logical; if TRUE, include the external elevation in the
hypsograph plot. Default is FALSE.}
}
\value{
ggplot object
}
\description{
Plot hypsograph
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
plot_hyps(aeme = aeme)
}
