% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_met_tile.R
\name{plot_met_tile}
\alias{plot_met_tile}
\title{Plot a tile plot of meteorological data}
\usage{
plot_met_tile(aeme, var_aeme = "MET_tmpair", use_hydro_year = TRUE, var_inp)
}
\arguments{
\item{aeme}{Aeme object.}

\item{var_aeme}{Character. Variable to plot. Can be one of:
\itemize{
\item \code{"MET_tmpair"}: Air temperature
\item \code{"MET_pprain"}: Rainfall
\item \code{"MET_wndspd"}: Wind speed
\item \code{"MET_humrel"}: Relative humidity
\item \code{"MET_radswd"}: Shortwave radiation
\item \code{"MET_radlwd"}: Longwave radiation
\item \code{"MET_pres"}: Atmospheric pressure
\item \code{"MET_ppsnow"}: Snowfall
\item \code{"MET_wnddir"}: Wind direction
}}

\item{use_hydro_year}{Logical. If \code{TRUE}, the hydrological year is used.
The hydrological year starts in October for the northern hemisphere and in
July for the southern hemisphere. If \code{FALSE}, the calendar year is used.}

\item{var_inp}{Character. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use
\code{var_aeme} instead.}
}
\value{
A ggplot object
}
\description{
Plot a tile plot of meteorological data
}
