% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phs.R
\name{plot_phs}
\alias{plot_phs}
\title{Plot phytoplankton variables}
\usage{
plot_phs(
  aeme,
  model,
  add_obs = TRUE,
  depth_range = NULL,
  remove_spin_up = TRUE,
  ens_n = 1
)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; model(s) to plot. If missing, all models
in the Aeme object will be plotted.}

\item{add_obs}{logical; add observations to plot}

\item{depth_range}{numeric; range of depths to plot. Default is NULL, which
averages over all depths.}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}

\item{ens_n}{integer; ensemble number to plot. Default is 1.}
}
\value{
A ggplot object
}
\description{
Plot phytoplankton variables
}
