% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot multi-variable timeseries}
\usage{
plot_ts(
  aeme,
  model,
  var_sim,
  remove_spin_up = TRUE,
  add_obs = TRUE,
  depth_range = NULL,
  ens_n = 1
)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; model(s) to plot. If missing, all models
in the Aeme object will be plotted.}

\item{var_sim}{character; variable in the AEME format (e.g. "HYD_temp"). Can
be a vector.}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}

\item{add_obs}{logical; add observations to the plot. Default is TRUE.}

\item{depth_range}{numeric; range of depths to average. Default is NULL,
which averages over all depths.}

\item{ens_n}{integer; ensemble number to plot. Default is 1.}
}
\value{
A ggplot object
}
\description{
Plot multi-variable timeseries
}
