% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbal_annual.R
\name{plot_wbal_annual}
\alias{plot_wbal_annual}
\title{Plot annual water balance components}
\usage{
plot_wbal_annual(aeme, model, lake_frac = FALSE, remove_spin_up = FALSE)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{character vector; model(s) to plot. If missing, all models
in the Aeme object will be plotted.}

\item{lake_frac}{Logical. If TRUE, water balance components are expressed as
a fraction of lake volume. Default is FALSE.}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}
}
\value{
ggplot2 object
}
\description{
Calculates and plots annual water balance components including evaporation,
precipitation, inflow, outflow, and net balance for each model in the Aeme
object.
}
