% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbal_comp.R
\name{plot_wbal_comp}
\alias{plot_wbal_comp}
\title{Plot Water Balance Components}
\usage{
plot_wbal_comp(aeme)
}
\arguments{
\item{aeme}{Aeme object with model output and observations added. Must have
model output \code{run_aeme()}}
}
\value{
ggplot object
}
\description{
Plot Water Balance Components
}
