% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbal_summaries.R
\name{plot_wbal_summaries}
\alias{plot_wbal_summaries}
\title{Plot Water Balance Summaries}
\usage{
plot_wbal_summaries(wbal)
}
\arguments{
\item{wbal}{Water balance object created with `get_wbal_components()`}
}
\value{
A ggplot2 object with water balance summaries
}
\description{
Plot Water Balance Summaries
}
