% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbal_ts.R
\name{plot_wbal_ts}
\alias{plot_wbal_ts}
\title{Plot Water Balance Components as Timeseries}
\usage{
plot_wbal_ts(
  aeme,
  var_aeme = c("LKE_lvlwtr", "LKE_inflow", "LKE_outflow", "LKE_pcpvol", "LKE_Qe",
    "HYD_surft"),
  add_model = TRUE
)
}
\arguments{
\item{aeme}{Aeme object with model output and observations added. Must have
model output \code{run_aeme()}}

\item{add_model}{logical; whether to overlay model output on the estimated
components. Defaults to TRUE.}

\item{vars}{character vector of AEME variable names to plot. Options are
\code{"LKE_lvlwtr"}, \code{"LKE_inflow"}, \code{"LKE_outflow"},
\code{"LKE_Qe"}, \code{"HYD_surft"}. Defaults to all.}
}
\value{
ggplot object
}
\description{
Plot Water Balance Components as Timeseries
}
