% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_status.R
\name{precip_status}
\alias{precip_status}
\title{Get current precipitation status in Aeme object}
\usage{
precip_status(aeme)
}
\arguments{
\item{aeme}{Aeme object.}
}
\value{
character. Either "met" if precipitation is set in meteorological
data, "inflow" if it is set as an inflow, or "none" if it is not set in
either.
}
\description{
This function checks whether precipitation is currently set as a meteorological input
or as an inflow in the Aeme object. It examines the meteorological data for
precipitation values and the inflow data for a precipitation inflow.
}
