% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_kw_regime.R
\name{print.kw_regime}
\alias{print.kw_regime}
\title{Print a Kw regime suggestion}
\usage{
\method{print}{kw_regime}(x, ...)
}
\arguments{
\item{x}{an object of class \code{kw_regime} (see
\code{\link[=suggest_kw_regime]{suggest_kw_regime()}}).}

\item{...}{unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Kw regime suggestion
}
