% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nml_helpers.R
\name{print.nml}
\alias{print.nml}
\title{Print an nml object}
\usage{
\method{print}{nml}(x, ...)
}
\arguments{
\item{x}{An nml object.}

\item{...}{Further arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Renders the nml object in GLM nml file format.
}
