% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{r_pearson}
\alias{r_pearson}
\title{Pearson Correlation Coefficient (r)}
\usage{
r_pearson(obs, sim, na.rm = TRUE)
}
\arguments{
\item{obs}{Numeric vector of observed values.}

\item{sim}{Numeric vector of simulated (modelled) values.}

\item{na.rm}{Logical; should missing values (NA) in either vector be removed
pairwise before computation? Default is \code{TRUE}.}
}
\value{
numeric; Pearson correlation coefficient ranging from -1 to 1.
}
\description{
Calculates the linear correlation between simulated and observed values.
}
