% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gotm_profile.R
\name{read_gotm_profile}
\alias{read_gotm_profile}
\title{Read GOTM profile file}
\usage{
read_gotm_profile(file)
}
\arguments{
\item{file}{Path to GOTM profile file}
}
\value{
Data frame with columns "datetime", "depth_m" and "value"
}
\description{
Read GOTM profile file
}
