% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_config.R
\name{read_model_config}
\alias{read_model_config}
\title{Read in model configuration files for a given model and lake directory}
\usage{
read_model_config(model, lake_dir)
}
\arguments{
\item{model}{character; model name ("dy_cd", "glm_aed", "gotm_wet"). Only
one model at a time.}

\item{path}{character; directory which contains the model configuration files.}
}
\value{
List with model configuration components. This includes a 'hydrodynamic'
list with hydrodynamic model configuration and a 'bgc' list with biogeochemistry
model configuration (if applicable).
}
\description{
Read in model configuration files for a given model and lake directory
}
