% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_hypsograph.R
\name{read_model_hypsograph}
\alias{read_model_hypsograph}
\title{Load model hypsograph from configuration}
\usage{
read_model_hypsograph(model, lake_dir, file = NULL)
}
\arguments{
\item{model}{Model name. One of "gotm_wet", "glm_aed", or "dy_cd".}

\item{lake_dir}{Directory of lake model outputs}
}
\value{
Dataframe of hypsograph with columns elev, area, and depth
}
\description{
Load model hypsograph from configuration
}
