% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nml.R
\name{read_nml}
\alias{read_nml}
\title{read in a GLM simulation *.nml file}
\usage{
read_nml(nml_file)
}
\arguments{
\item{nml_file}{a string with the path to the GLM glm2.nml file, or}
}
\value{
glm_nml a nml (a list) for GLM config
}
\description{
read in a GLM simulation *.nml file and create a list.  \cr
}
\seealso{
\link{get_nml_value}
 @source glmtools package: https://github.com/GLEON/glmtools
}
\author{
Jordan S. Read
}
\keyword{methods}
