% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_inflow.R
\name{remove_inflow}
\alias{remove_inflow}
\title{Remove inflow data from Aeme object}
\usage{
remove_inflow(aeme, inflow_id = NULL, all = FALSE)
}
\arguments{
\item{aeme}{aeme; object.}

\item{inflow_id}{A string specifying the identifier for the inflow.}

\item{all}{A logical value indicating whether to remove all inflows. If TRUE,
the inflow_id argument is ignored and all inflows are removed.}
}
\value{
Aeme object with inflow removed
}
\description{
Remove inflow data from an Aeme object. You can specify a particular inflow 
to remove by providing its identifier, or you can remove all inflows at once. 
If you choose to remove all inflows, the inflow_id argument will be ignored.
}
