% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_param.R
\name{remove_param}
\alias{remove_param}
\title{Remove parameter(s) from Aeme object}
\usage{
remove_param(aeme, name, model, file)
}
\arguments{
\item{aeme}{Aeme object.}

\item{name}{character vector with names of parameters to remove. If missing,
all parameters (within \code{model}/\code{file}, if supplied) are removed.}

\item{model}{character vector; only remove parameters for these models.
If missing, parameters are removed regardless of model.}

\item{file}{character vector; only remove parameters in these files. If
missing, parameters are removed regardless of file.}
}
\value{
Aeme object with parameters removed
}
\description{
Remove parameter(s) from Aeme object
}
