% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_modelvars.R
\name{rename_modelvars}
\alias{rename_modelvars}
\title{Rename variables among the models and controlled vocabulary}
\usage{
rename_modelvars(
  input,
  type_output = "name_parse",
  type_input = "name",
  verbose = F
)
}
\arguments{
\item{input}{string; vector of names}

\item{type_output}{string; type of output}

\item{type_input}{string; type of input}

\item{verbose}{boolean; print changes to console}
}
\value{
vector of changed names
}
\description{
Rename variables among the models and controlled vocabulary
}
