% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_wbal_param.R
\name{reset_wbal_param}
\alias{reset_wbal_param}
\title{Reset water balance parameters}
\usage{
reset_wbal_param(aeme)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
Aeme object with water balance parameters reset
}
\description{
This function resets the water balance parameters in the Aeme object. This is
useful if you want to start fresh with a new set of parameters for example 
if you add/remove a inflow or change the meteorological data.
}
