% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_params.R
\name{retrieve_params}
\alias{retrieve_params}
\title{Retrieve parameters for each model}
\usage{
retrieve_params(model, module, par)
}
\arguments{
\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{module}{character; vector of modules to retrieve parameters for.}

\item{par}{character; vector of parameters to retrieve.}
}
\value{
dataframe; of parameters filtered by model, module and parameter.
}
\description{
Retrieve parameters for each model
}
