% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ensemble.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run ensemble}
\usage{
run_ensemble(
  config,
  model,
  verbose = FALSE,
  debug = FALSE,
  timeout = 0,
  parallel = FALSE,
  dir = "."
)
}
\arguments{
\item{parallel}{}
}
\value{
Runs the model
}
\description{
Run ensemble
}
