% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_water_balance.R
\name{sat_vapour_pressure}
\alias{sat_vapour_pressure}
\title{Saturation vapour pressure}
\usage{
sat_vapour_pressure(Ts)
}
\arguments{
\item{Ts}{Numeric. Water surface temperature (°C).}
}
\value{
Numeric. Saturation vapour pressure (hPa).
}
\description{
Calculates saturation vapour pressure at the water surface using the
Magnus formula.
}
\examples{
sat_vapour_pressure(20)
sat_vapour_pressure(c(15, 20, 25))
}
