% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{show,Aeme-method}
\alias{show,Aeme-method}
\title{Print Aeme object to the console}
\usage{
\S4method{show}{Aeme}(object)
}
\arguments{
\item{object}{An Aeme object.}
}
\value{
prints the Aeme object to the console.
}
\description{
Show an Aeme object in the console
}
\details{
This method prints the Aeme output in a readable format to the console.
}
