% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simstrat_aed2_parameters}
\alias{simstrat_aed2_parameters}
\title{Example dataframe used for calibrating the Simstrat-AED2 model.}
\format{
\code{simstrat_aed2_parameters}
A data frame with columns:
\describe{
\item{model}{Model for the parameter}
\item{file}{File in which the parameter is stored}
\item{name}{Name of the parameter}
\item{value}{Value of the parameter}
\item{min}{Minimum range of the parameter}
\item{max}{Maximum range of the parameter}
\item{module}{Whether the parameter belongs to the "hydrodynamic" or
"bgc" (biogeochemical) part of the model}
\item{par}{Short parameter name}
}
}
\source{
Package development.
}
\usage{
simstrat_aed2_parameters
}
\description{
Physical (hydrodynamic) parameters from the Simstrat \code{.par} file and a
subset of biogeochemical initial-concentration parameters from the
\code{aed2.nml} file (oxygen, carbon, silica, nitrogen, phosphorus, and organic
matter modules). This has the values in the default template files and
50 \% parameter ranges for sensitivity analysis.
}
\keyword{datasets}
