% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{summary,aeme-method}
\alias{summary,aeme-method}
\title{Summarise an aeme object}
\usage{
\S4method{summary}{aeme}(object, ...)
}
\arguments{
\item{object}{An aeme object.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
prints the aeme object to the console.
}
\description{
Summarise an aeme object
}
\details{
This method summarises the aeme output.
}
