% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{time,Aeme-method}
\alias{time,Aeme-method}
\title{Access time slot}
\usage{
\S4method{time}{Aeme}(aeme)
}
\arguments{
\item{aeme}{A Aeme object.}
}
\value{
list of time characteristics
}
\description{
Access time slot
}
