% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{time<-,aeme-method}
\alias{time<-,aeme-method}
\title{Set time in aeme object}
\usage{
\S4method{time}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New time data to be assigned.}
}
\value{
A modified aeme object with updated time slot.
}
\description{
Update the time slot of an aeme object
}
\details{
This method updates the "time" slot of a aeme object with new data.
}
