% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{time<-}
\alias{time<-}
\alias{time<-,Aeme-method}
\title{Set the time slot of an Aeme object}
\usage{
time(aeme) <- value

\S4method{time}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New time list to assign.}
}
\value{
A modified Aeme object with updated time slot.
}
\description{
Set the time slot of an Aeme object
}
