% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_init.R
\name{update_init}
\alias{update_init}
\title{Update initial conditions in AEME object based on lake observations.}
\usage{
update_init(aeme, model_controls = NULL)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model_controls}{data.frame; model configuration, typically loaded via
\code{\link[=get_model_controls]{get_model_controls()}}.}
}
\value{
Aeme object with updated initial conditions
}
\description{
Update initial conditions in AEME object based on lake observations.
}
