% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{update_output,aeme-method}
\alias{update_output,aeme-method}
\title{Update Output Method}
\usage{
\S4method{update_output}{aeme}(aeme_data, new_output)
}
\arguments{
\item{aeme_data}{A aeme object.}

\item{new_output}{New output data to be assigned.}
}
\value{
A modified aeme object with updated output slot.
}
\description{
Update the output slot of an aeme object
}
\details{
This method updates the "output" slot of a aeme object with new data.
}
