% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{water_balance,aeme-method}
\alias{water_balance,aeme-method}
\title{Access water_balance slot}
\usage{
\S4method{water_balance}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of water_balance characteristics
}
\description{
Access water_balance slot
}
