% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{water_balance<-,aeme-method}
\alias{water_balance<-,aeme-method}
\title{Set water_balance in aeme object}
\usage{
\S4method{water_balance}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New water_balance data to be assigned.}
}
\value{
A modified aeme object with updated water_balance slot.
}
\description{
Update the water_balance slot of an aeme object
}
\details{
This method updates the "water_balance" slot of a aeme object with new data.
}
