% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{water_balance<-}
\alias{water_balance<-}
\alias{water_balance<-,Aeme-method}
\title{Set the water_balance slot of an Aeme object}
\usage{
water_balance(aeme) <- value

\S4method{water_balance}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New water_balance list to assign.}
}
\value{
A modified Aeme object with updated water_balance slot.
}
\description{
Set the water_balance slot of an Aeme object
}
