% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hypsograph.R
\name{add_hypsograph}
\alias{add_hypsograph}
\title{Add hypsograph to Aeme object}
\usage{
add_hypsograph(aeme = NULL, hypsograph = NULL, surf_elev = 0, ext_elev = 0)
}
\arguments{
\item{aeme}{aeme; object.}

\item{hypsograph}{data frame with columns "depth" and "area". Depth should be
monotonic decreasing and area should be monotonic increasing. If elevation is
not provided, it will be calculated as depth + lake elevation.}

\item{ext_elev}{numeric; metres to extend the hypograph by.}
}
\value{
Aeme object with hypsograph added
}
\description{
Add hypsograph to Aeme object
}
\examples{
depth <- seq(-5, 0, by = 0.5)
area <- c(0, 10, 30, 50, 70, 90, 110, 130, 150, 170, 190)
hypsograph <- data.frame(depth = depth, area = area)
}
