% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hypsograph.R
\name{add_hypsograph}
\alias{add_hypsograph}
\title{Add hypsograph to Aeme object}
\usage{
add_hypsograph(
  aeme = NULL,
  hypsograph = NULL,
  surf_elev = 0,
  lake_depth = NULL,
  lake_area = NULL,
  ext_elev = 0
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{hypsograph}{data frame with columns "depth" and "area". Depth should be
monotonic decreasing and area should be monotonic increasing. If elevation is
not provided, it will be calculated as depth + lake elevation.
If NULL, a simple hypsograph will be generated using lake depth and area.}

\item{surf_elev}{Lake surface elevation (m). Required if hypsograph is NULL.}

\item{lake_depth}{Lake maximum depth (m). Required if hypsograph is NULL.}

\item{lake_area}{Lake surface area (m2). Required if hypsograph is NULL.}

\item{ext_elev}{numeric; metres to extend the hypograph by.}
}
\value{
Aeme object with hypsograph added
}
\description{
Add hypsograph to Aeme object
}
\examples{
depth <- seq(-5, 0, by = 0.5)
area <- c(0, 10, 30, 50, 70, 90, 110, 130, 150, 170, 190)
hypsograph <- data.frame(depth = depth, area = area)
}
