% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_inflow.R
\name{add_inflow}
\alias{add_inflow}
\title{Set inflow data from Aeme object}
\usage{
add_inflow(aeme, inflow, inflow_id)
}
\arguments{
\item{aeme}{aeme; object.}

\item{inflow}{A data frame containing inflow data. Must include columns "Date",
"HYD_flow", "HYD_temp", and "CHM_salt". 
An "inflow_id" column can be included to identify the inflow or it can be
provided separately. If multiple unique inflow_ids are present, an error
will be raised.}

\item{inflow_id}{A string specifying the identifier for the inflow. If not
provided, the function will look for an "inflow_id" column in the inflow
data frame. If the column is present but contains multiple unique values, an
error will be raised.}
}
\value{
Aeme object with inflow added
}
\description{
Replace existing inflow data in an Aeme object with new inflow data. The new
inflow data can be provided as either a named list of data frames or a single
data frame containing an "inflow_id" column to identify different inflows.
If the original inflows include a "precip" inflow, it will be
retained in the updated inflows.
}
