% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_met.R
\name{add_met}
\alias{add_met}
\title{Add meteorological data to Aeme object}
\usage{
add_met(aeme, met)
}
\arguments{
\item{aeme}{Aeme object.}

\item{met}{data frame with meteorological data. Must include columns "Date",
"MET_radswd", "MET_radswd", "MET_pprain" and "MET_wndspd" or "MET_wnduvu" and
"MET_wnduvv".}
}
\value{
Aeme object with meteorological data added
}
\description{
Add meteorological data to Aeme object
}
