% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_output.R
\name{add_output}
\alias{add_output}
\title{Add model output to Aeme object}
\usage{
add_output(aeme, out)
}
\arguments{
\item{aeme}{Aeme object}

\item{out}{list with model output loaded with `load_output()`}
}
\value{
Aeme object with model output added
}
\description{
Add model output to Aeme object
}
