% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\docType{class}
\name{Aeme}
\alias{Aeme}
\alias{Aeme-class}
\title{Aeme Class}
\description{
S4 Class representing AEME data
}
\details{
This class represents data related to a lake for running AEME. Items in bold
are required to run the models.
}
\section{Slots}{

\describe{
\item{\code{lake}}{A list representing lake information. \itemize{
\item \strong{\code{name}}: character; lake name.
\item \strong{\code{id}}: character; lake identifier.
\item \strong{\code{latitude}}: numeric; lake latitude.
\item \strong{\code{longitude}}: numeric; lake longitude.
\item \strong{\code{elevation}}: numeric; lake elevation.
\item \strong{\code{depth}}: numeric; lake depth.
\item \strong{\code{area}}: numeric; lake area.
}}

\item{\code{time}}{A list representing time information. \itemize{
\item \strong{\code{start}}: character; start date.
\item \strong{\code{stop}}: character; end date.
\item \strong{\code{timestep}}: numeric; time step.
\item \strong{\code{spin_up}}: list; spin up information for each model
}}

\item{\code{configuration}}{A list representing each model's configuration. \itemize{
\item \code{model_controls}: dataframe; Model controls for simulation.
\item \code{dy_cd}: list; DYRESM-CAEDYM configuration.
\item \code{glm_aed}: list; GLM-AED configuration.
\item \code{gotm_wet}: list; GOTM-WET configuration.
}}

\item{\code{observations}}{A list representing observation information. \itemize{
\item \code{lake}: dataframe; lake observations.
\item \code{level}: dataframe; lake level observations.
}}

\item{\code{input}}{A list representing input information. \itemize{
\item \code{init_profile}: dataframe; initial temperature profile (if none
use NULL or leave empty; if empty/NULL, the observations file will be used).
\item \strong{\code{init_depth}}: numeric; initial height of lake surface relative to
the bottom (m).
\item \strong{\code{hypsograph}}: dataframe; hypsograph.
\item \strong{\code{meteo}}: dataframe; meteorological data.
\item \strong{\code{use_lw}}: logical; use longwave radiation.
\item \strong{\code{Kw}}: numeric; light extinction coefficient (m-1).
}}

\item{\code{inflows}}{A list representing inflows information. \itemize{
\item \code{data}: named list of inflow dataframes.
\item \code{factor}: named list; inflow factors for each model.
}}

\item{\code{outflows}}{A list representing outflows information. \itemize{
\item \code{data}: named list of outflow dataframes.
\item \code{factor}: named list; outflow factors for each model.
\item \code{lvl}: numeric; height of lake level outflow.
}}

\item{\code{water_balance}}{A list representing water balance information. \itemize{
\item\strong{\code{method}}: integer; Method for calculating water balance.
1 = none, 2 = outflows, 3 = inflows and outflows.
\item\strong{\code{use}}: character; Can be 'obs' or 'mod'. Use observations
or modelled data for water balance.
\item{\code{data}}: list of dataframe for water balance.
}}

\item{\code{output}}{A list representing output information. \itemize{
\item \code{dy_cd}: list; DYRESM-CAEDYM output.
\item \code{glm_aed}: list; GLM-AED output.
\item \code{gotm_wet}: list; GOTM-WET output.
}}

\item{\code{parameters}}{A dataframe representing model parameters.}
}}

