% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\docType{class}
\name{aeme}
\alias{aeme}
\alias{aeme-class}
\title{aeme Class}
\description{
S4 Class representing AEME data
}
\details{
This class represents data related to a lake for running AEME.
}
\section{Slots}{

\describe{
\item{\code{lake}}{A list representing lake information.}

\item{\code{catchment}}{A list representing catchment information.}

\item{\code{time}}{A list representing time information.}

\item{\code{observations}}{A list representing observation information.}

\item{\code{input}}{A list representing input information.}

\item{\code{inflows}}{A list representing inflows information.}

\item{\code{outflows}}{A list representing outflows information.}

\item{\code{output}}{A list representing output information.}
}}

