% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\docType{class}
\name{aeme}
\alias{aeme}
\alias{aeme-class}
\title{aeme Class}
\usage{
aeme(lake, catchment, time, observations, input, inflows, outflows, output)
}
\arguments{
\item{lake}{List representing lake information.}

\item{catchment}{List representing catchment information.}

\item{time}{List representing time information.}

\item{observations}{List representing observation information.}

\item{input}{List representing input information.}

\item{inflows}{List representing inflows information.}

\item{outflows}{List representing outflows information.}

\item{output}{List representing output information.}
}
\value{
An instance of the aeme class.
}
\description{
S4 Class representing AEME data
}
\details{
This class represents data related to a lake.
}
\section{Slots}{

\describe{
\item{\code{lake}}{A list representing lake information.}

\item{\code{catchment}}{A list representing catchment information.}

\item{\code{time}}{A list representing time information.}

\item{\code{observations}}{A list representing observation information.}

\item{\code{input}}{A list representing input information.}

\item{\code{inflows}}{A list representing inflows information.}

\item{\code{outflows}}{A list representing outflows information.}

\item{\code{output}}{A list representing output information.}
}}

