% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{aeme_constructor}
\alias{aeme_constructor}
\title{Constructor function for aeme class}
\usage{
aeme_constructor(
  lake,
  catchment,
  time,
  observations,
  input,
  inflows,
  outflows,
  output
)
}
\arguments{
\item{lake}{List representing lake information.}

\item{catchment}{List representing catchment information.}

\item{time}{List representing time information.}

\item{observations}{List representing observation information.}

\item{input}{List representing input information.}

\item{inflows}{List representing inflows information.}

\item{outflows}{List representing outflows information.}

\item{output}{List representing output information.}
}
\value{
An instance of the aeme class.
}
\description{
Constructor function for aeme class
}
