% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{aeme_constructor}
\alias{aeme_constructor}
\title{Constructor function for Aeme class}
\usage{
aeme_constructor(
  lake,
  time,
  configuration,
  observations,
  input,
  inflows,
  outflows,
  water_balance,
  output,
  parameters,
  print = TRUE
)
}
\arguments{
\item{lake}{List representing lake information.}

\item{time}{List representing time information.}

\item{configuration}{List representing configuration information.}

\item{observations}{List representing observation information.}

\item{input}{List representing input information.}

\item{inflows}{List representing inflows information.}

\item{outflows}{List representing outflows information.}

\item{water_balance}{List representing water balance information.}

\item{output}{List representing output information.}

\item{parameters}{Dataframe containing model parameters.}

\item{print}{Logical; print messages. Default is TRUE.}
}
\value{
An instance of the Aeme class.
}
\description{
Constructor function for Aeme class
}
