% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aeme_parameters}
\alias{aeme_parameters}
\title{Example dataframe used for calibrating AEME models.}
\format{
\subsection{\code{aeme_parameters}}{

A data frame with 34 rows and 10 columns:
\describe{
\item{model}{Model for the parameter}
\item{file}{File in which the parameter is stored}
\item{name}{Name of the parameter}
\item{value}{Value of the parameter}
\item{min}{Minimum range of the parameter}
\item{max}{Maximum range of the parameter}
\item{group}{Phytoplankton Group for the parameter, only applies to
phytoplankton parameters}
\item{index}{Index for the parameter in the model file, only used for
parameters that have multiple values in a vector such as
"sediment/sed_temp_mean" in GLM-AED}
\item{module}{Module for the parameter in the model, useful to help
identify parameters}
\item{var_sim}{Variable name in AEME that it influences, useful to help identify parameters}
}
}
}
\source{
Package development.
}
\usage{
aeme_parameters
}
\description{
An example dataframe used for inputting and calibrating AEME models.
}
\keyword{datasets}
