% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme_to_inflow.R
\name{aeme_to_inflow}
\alias{aeme_to_inflow}
\title{Convert AEME model outputs to inflow format}
\usage{
aeme_to_inflow(aeme)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
List of data frames with inflow variables for each model
}
\description{
Convert AEME model outputs to inflow format
}
