% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_depth_data.R
\name{align_depth_data}
\alias{align_depth_data}
\title{Align observation depth data with model data}
\usage{
align_depth_data(aeme, model, var_sim, ens_n = 1, return_df = TRUE)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be \code{dy_cd}, \code{glm_aed},
\code{gotm_wet}.}

\item{var_sim}{character; variable in the AEME format (e.g. "HYD_temp").}

\item{ens_n}{integer; ensemble number to plot. Default is 1.}

\item{return_df}{logical; if TRUE, return a dataframe; if FALSE, return a
list. Default is TRUE.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{Date}: Date of observation
\item \code{depth}: Depth of observation
\item \code{elev}: Elevation of observation
\item \code{Model}: Model name
\item \code{var_sim}: Variable name
\item \code{value}: Value of the variable
\item \code{depth_from}: Depth from which the variable is extracted
\item \code{depth_to}: Depth to which the variable is extracted
}
}
\description{
Align observation depth data with model data
}
