% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_model.R
\name{assess_model}
\alias{assess_model}
\title{Assess model performance}
\usage{
assess_model(aeme_data, model, var_sim = "HYD_temp")
}
\arguments{
\item{aeme_data}{aeme; data object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}
}
\value{
data frame with model performance statistics for each model and
variable
}
\description{
Assess model performance
}
