% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_model.R
\name{assess_model}
\alias{assess_model}
\title{Assess model performance}
\usage{
assess_model(aeme_data, model, var_sim = "HYD_temp")
}
\arguments{
\item{aeme_data}{aeme; data object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{var_sim}{string; of variable to plot}
}
\value{
Data frame with model performance statistics for each model and
variable. These include:
\itemize{
\item{bias}{ - Bias}
\item{mae}{ - Mean absolute error}
\item{rmse}{ - Root mean square error}
\item{nmae}{ - Normalised mean absolute error}
\item{nse}{ - Nash-Sutcliffe efficiency}
\item{d2}{ - Index of agreement model skill score Willmott index}
\item{r}{ - Pearson correlation coefficient}
\item{rs}{ - Spearman correlation coefficient}
\item{r2}{ - R-squared value from linear model}
\item{B}{ - Bardsley coefficient}
\item{n}{ - number of observations}
}
}
\description{
Assess model performance
}
