% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ensemble.R
\name{build_ensemble}
\alias{build_ensemble}
\title{Build model configuration directories}
\usage{
build_ensemble(
  aeme_data = NULL,
  config = NULL,
  model = c("dy_cd", "glm_aed", "gotm_wet"),
  mod_ctrls,
  inf_factor = c(glm_aed = 1, dy_cd = 1, gotm_wet = 1),
  outf_factor = c(glm_aed = 1, dy_cd = 1, gotm_wet = 1),
  ext_elev = 0,
  use_bgc = TRUE,
  use_lw = FALSE,
  hum_type = 3,
  dir = "."
)
}
\arguments{
\item{aeme_data}{aeme; data object.}

\item{config}{list; loaded via `config <- yaml::read_yaml("aeme.yaml")`}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{mod_ctrls}{dataframe; of configuration loaded from
"mod_ctrls.csv".}

\item{inf_factor}{vector; containing numeric factor to multiple the inflows.
Needs to be named according to the model.}

\item{outf_factor}{vector; containing numeric factor to multiple the
outflows. Needs to be named according to the model.}

\item{ext_elev}{numeric; extension in elevation for the hypsogrph in metres.}

\item{use_bgc}{boolean; switch to use the biogeochemical model.}

\item{use_lw}{boolean; use incoming longwave radiation. Only applies to
GLM-AED.}

\item{hum_type}{numeric; GOTM humidity metric [1=relative humidity (%),
2=wet-bulb temperature, 3=dew point temperature, 4=specific humidity (kg/kg)]
Default = 3.}

\item{dir}{filepath; where input files are located relative to `config`.}
}
\value{
builds the model ensemble configuration.
}
\description{
Configure an ensemble of lake model simulations from basic set of inputs.
}
\examples{
install.packages("configr", repos = "http://cran.us.r-project.org")
tmpdir <- tempdir()
aeme_dir <- system.file("extdata/lake/", package = "AEME")
# Copy files from package into tempdir
file.copy(aeme_dir, tmpdir, recursive = TRUE)
dir <- file.path(tmpdir, "lake")
aeme_data <- yaml_to_aeme(dir = dir, "aeme.yaml")
mod_ctrls <- read.csv(file.path(dir, "model_controls.csv"))
inf_factor = c("glm_aed" = 1)
outf_factor = c("glm_aed" = 1)
model <- c("glm_aed")
build_ensemble(dir = dir, aeme_data = aeme_data, model = model,
               mod_ctrls = mod_ctrls, inf_factor = inf_factor, ext_elev = 5,
               use_bgc = FALSE, use_lw = TRUE)
}
