% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ensemble.R
\name{build_ensemble}
\alias{build_ensemble}
\title{Configure an ensemble of lake model simulations from basic set of inputs}
\usage{
build_ensemble(
  config,
  model = c("dy_cd", "glm_aed", "gotm_pclake"),
  spinups = c(glm_aed = 0, dy_cd = 0, gotm_pclake = 365, gotm_wet = 365),
  key.cfg = data.frame(readxl::read_excel("data/config_files/key_ensemble_Default.xlsx")),
  key_naming,
  inf.factor = c(glm_aed = 1, dy_cd = 1, gotm_wet = 1),
  wdr.factor = c(glm_aed = 1, dy_cd = 1, gotm_wet = 1),
  ext_elev = 0,
  config_dir = "data/config_files/",
  use_bgc = TRUE,
  use_lw = FALSE,
  hum_type = 3,
  dir = "."
)
}
\arguments{
\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{spinups}{vector; containing the number of days in the simulation to be
allocated to a spin-up period. Needs to be named using the models.}

\item{key.cfg}{dataframe; of configuration loaded from
"key_ensemble_Default.xlsx".}

\item{key_naming}{dataframe; of corresponding names for the different models.}

\item{inf.factor}{vector; containing factor to multiple the inflows. Needs to
be named according to the model.}

\item{wdr.factor}{vector; containing factor to multiple the outflows. Needs to
be named according to the model.}

\item{ext_elev}{numeric; extension in elevation for the hypsogrph in (m).}

\item{config_dir}{filepath; to directory which contains the model
configuration files.}

\item{use_bgc}{boolean; switch to use the biogeochemical model.}

\item{lid}{character; of ID number used for the lake}

\item{lake}{sf polygon of the lake shoreline.}

\item{date.range}{vector; of start and end dates}

\item{hyps}{dataframe; of the hypsograph. Required columns are `elev` and
`area`.}

\item{lvl}{dataframe; of average surface water level or time-series of daily
water level (masl).}

\item{met}{dataframe of meteorological data.}

\item{inf}{list; of dataframe containing inflows. Each inflow should be named
in the list.}

\item{wdr}{dataframe; containing outflow for the lake. Date plus a col for
each outflow.}
}
\value{
builds the model ensemble configuration.
}
\description{
Configure an ensemble of lake model simulations from basic set of inputs
}
