% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fairall.R
\name{calc_fairall}
\alias{calc_fairall}
\title{Calculate Fairall et al 1996 bulk fluxes}
\usage{
calc_fairall(
  sst,
  airt,
  u10,
  v10,
  airp,
  hum,
  precip,
  rain_impact = TRUE,
  calc_evaporation = TRUE
)
}
\arguments{
\item{sst}{Sea surface temperature [degC]}

\item{airt}{Air temperature [degC]}

\item{u10}{Eastward wind speed at 10m [m/s]}

\item{v10}{Northward wind speed at 10m [m/s]}

\item{airp}{Air pressure [Pa]}

\item{hum}{Relative humidity [%]}

\item{precip}{Precipitation rate [m/s]}

\item{rain_impact}{Logical, should rain impact be included? Default is TRUE}

\item{calc_evaporation}{Logical, should evaporation be calculated? Default is TRUE}
}
\value{
List with the following elements:
\itemize{
 \item{tau_x}{Eastward wind stress [N/m^2]}
 \item{tau_y}{Northward wind stress [N/m^2]}
 \item{h}{Sensible heat flux [W/m^2]}
 \item{e}{Latent heat flux [W/m^2]}
 \item{evap}{Evaporation rate [m/s]}
 }

 @noRd
}
\description{
Calculate airsea fluxes using Fairall et al 1996
}
