% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catchment_yields}
\alias{catchment_yields}
\title{Catchment yields}
\format{
## `catchment_yields`
A data frame with 4 rows and 4 columns:
\describe{
  \item{Type}{Type of catchment}
  \item{TN_kg/ha/y}{Total nitrogen yield per hectare per year.}
  \item{TP_kg/ha/y}{Total phosphorus yield per hectare per year.}
  \item{TSS_kg/ha/y}{Total suspended sediment yield per hectare per year.}
}
}
\source{
Package development.
}
\usage{
catchment_yields
}
\description{
A reference table for catchment nutrient and sediment yields for calculating
inflow nutrients in AEME.
}
\author{
Chris McBride
}
\keyword{datasets}
