% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_aeme}
\alias{check_aeme}
\title{Check if object is a valid Aeme object}
\usage{
check_aeme(aeme)
}
\arguments{
\item{aeme}{object to check}
}
\value{
Invisibly returns the Aeme object if valid, otherwise throws an 
error.
}
\description{
Check if object is a valid Aeme object
}
