% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_AEME.R
\name{check_AEME}
\alias{check_AEME}
\title{Check if the package is working correctly}
\usage{
check_AEME()
}
\value{
TRUE if the package is working correctly
}
\description{
Check if the package is working correctly
}
