% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gotm_yaml.R
\name{check_gotm_yaml}
\alias{check_gotm_yaml}
\title{Check GOTM YAML configuration file for common issues}
\usage{
check_gotm_yaml(file)
}
\arguments{
\item{file}{Path to GOTM YAML configuration file}
}
\value{
Invisibly returns TRUE if no issues are found; otherwise, throws an 
error
}
\description{
Check GOTM YAML configuration file for common issues
}
