% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check model name and return standardized code}
\usage{
check_model(model)
}
\arguments{
\item{model}{Character vector of model names. Valid options are:
"DYRESM-CAEDYM", "GLM-AED", "GOTM-WET" or their corresponding codes
"dy_cd", "glm_aed", "gotm_wet".}
}
\value{
Character vector of standardized model codes.
}
\description{
Check model name and return standardized code
}
\examples{
check_model(c("GLM-AED", "gotm_wet"))
}
