% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_output.R
\name{check_model_output}
\alias{check_model_output}
\title{Check model output}
\usage{
check_model_output(aeme, model, path)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be \code{dy_cd}, \code{glm_aed},
\code{gotm_wet}.}

\item{path}{filepath; where input files are located relative to the current
working directory.}
}
\value{
Invisibly TRUE if model output passes checks; otherwise aborts
}
\description{
Check model output
}
