% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_output.R
\name{check_model_output}
\alias{check_model_output}
\title{Check model output}
\usage{
check_model_output(aeme, model, path)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}
}
\value{
Invisibly TRUE if model output passes checks; otherwise aborts
}
\description{
Check model output
}
