% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_simstrat_par.R
\name{check_simstrat_par}
\alias{check_simstrat_par}
\title{Check Simstrat par file for common issues}
\usage{
check_simstrat_par(file)
}
\arguments{
\item{file}{path to Simstrat \code{.par} (JSON) file}
}
\value{
Invisibly returns TRUE if no issues found, otherwise aborts with
informative messages
}
\description{
Check Simstrat par file for common issues
}
