% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_var_in_output.R
\name{check_var_in_output}
\alias{check_var_in_output}
\title{Check if a variable is present in the model output}
\usage{
check_var_in_output(aeme, model, var_sim, ens_n = 1)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{var_sim}{string; of variable to plot}

\item{ens_n}{numeric; ensemble number}
}
\value{
A list with logical matrices indicating if each variable is present per model,
plus an overall \code{all_present} flag.
}
\description{
Check if a variable is present in the model output
}
