% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_var_in_output.R
\name{check_var_in_output}
\alias{check_var_in_output}
\title{Check if a variable is present in the model output}
\usage{
check_var_in_output(aeme, model, var_sim, ens_n = 1)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}
}
\value{
A logical matrix indicating if the variable is present in the model
}
\description{
Check if a variable is present in the model output
}
