% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{cli_safe}
\alias{cli_safe}
\title{Inform messages respecting the global AEME.inform option}
\usage{
cli_safe(..., FUN = cli::cli_bullets, indent = TRUE, .envir = parent.frame())
}
\arguments{
\item{...}{arguments passed to cli_inform_safe()}

\item{FUN}{function to use for messaging, default is cli::cli_inform}

\item{indent}{logical, whether to indent the message, default is FALSE}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
Inform messages respecting the global AEME.inform option
}
