% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{configuration,Aeme-method}
\alias{configuration,Aeme-method}
\title{Access configuration slot}
\usage{
\S4method{configuration}{Aeme}(aeme)
}
\arguments{
\item{aeme}{A Aeme object.}
}
\value{
list of configuration characteristics
}
\description{
Access configuration slot
}
