% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{configuration<-,aeme-method}
\alias{configuration<-,aeme-method}
\title{Set configuration in aeme object}
\usage{
\S4method{configuration}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New configuration data to be assigned.}
}
\value{
A modified aeme object with updated configuration slot.
}
\description{
Update the configuration slot of an aeme object
}
\details{
This method updates the "configuration" slot of a aeme object with new data.
}
