% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{configuration<-}
\alias{configuration<-}
\alias{configuration<-,Aeme-method}
\title{Set the configuration slot of an Aeme object}
\usage{
configuration(aeme) <- value

\S4method{configuration}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New configuration list to assign.}
}
\value{
A modified Aeme object with updated configuration slot.
}
\description{
Set the configuration slot of an Aeme object
}
