% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{configuration<-}
\alias{configuration<-}
\title{Set configuration in Aeme object}
\usage{
configuration(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New configuration data to be assigned.}
}
\value{
A modified Aeme object with updated configuration slot.
}
\description{
Update the configuration slot of an Aeme object
}
