% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{configuration}
\alias{configuration}
\title{Access configuration slot}
\usage{
configuration(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of configuration characteristics
}
\description{
Access configuration slot
}
