% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dy_cd_parameters}
\alias{dy_cd_parameters}
\title{Example dataset of parameters for the DYRESM-CAEDYM model}
\format{
A data frame with 253 rows and 9 columns:
\describe{
  \item{model}{Model to which the parameter belongs}
  \item{file}{File in which the parameter is stored}
  \item{name}{Name of the parameter}
  \item{value}{Default value of the parameter}
  \item{min}{Minimum range of the parameter}
  \item{max}{Maximum range of the parameter}
  \item{group}{Phytoplankton group for the parameter; only applies to
    phytoplankton parameters}
  \item{index}{Index for parameters with multiple values in a vector (e.g.,
    `"sediment/sed_temp_mean"` in GLM-AED)}
  \item{module}{Model module associated with the parameter, useful for
    identifying functional groupings}
}
}
\source{
Created during package development.
}
\usage{
dy_cd_parameters
}
\description{
This dataset contains all parameters defined in the `.par` and `.cfg` file.
The values represent defaults from the standard configuration file, with 
±25% parameter ranges for use in sensitivity analysis or model calibration.
}
\keyword{datasets}
