% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_met.R
\name{expand_met}
\alias{expand_met}
\title{Expand a minimal set of meteorology inputs to a complete set of variables
suitable for all models}
\usage{
expand_met(met, lat, lon, elev, print.plot = FALSE)
}
\arguments{
\item{met}{data frame of source meteorology variables}

\item{lat}{numeric; latitude}

\item{lon}{numeric; longitude}

\item{elev}{numeric; elevation}

\item{print.plot}{logical; print the plot}
}
\value{
dataframe with expanded met variables
}
\description{
Expand a minimal set of meteorology inputs to a complete set of variables
suitable for all models
}
