% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrap_hyps.R
\name{extrap_hyps}
\alias{extrap_hyps}
\title{Extend hypsmetry to a greater elevation using linear extrapolation}
\usage{
extrap_hyps(hyps, z_range = 0.2, ext_elev)
}
\arguments{
\item{hyps}{dataframe; with hypsograph}

\item{z_range}{numeric; 0-1, representing fraction of hypsograph to be used.
Default is 0.2, which uses the top 20% of the hypsograph for extrapolation.
when extrapolating the hypsmetry.}

\item{ext_elev}{numeric; new max depth.}
}
\value{
dataframe with extrapolated hypsmetry
}
\description{
Extend hypsmetry to a greater elevation using linear extrapolation
}
