% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrap_hyps.R
\name{extrap_hyps}
\alias{extrap_hyps}
\title{Extend hypsometry to a greater elevation using linear extrapolation}
\usage{
extrap_hyps(hypsograph, z_range = 0.2, ext_elev)
}
\arguments{
\item{hypsograph}{dataframe; with hypsograph}

\item{z_range}{numeric; 0-1, representing fraction of hypsograph to be used
for extrapolation. Default is 0.2, which uses the top 20% of the hypsograph
for extrapolation.}

\item{ext_elev}{numeric; metres to extend the hypograph by.}
}
\value{
dataframe with extrapolated hypsmetry
}
\description{
Extend hypsometry to a greater elevation using linear extrapolation
}
