% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aed_sed_const2d_param.R
\name{get_aed_sed_const2d_param}
\alias{get_aed_sed_const2d_param}
\title{Get AED sed_const2d parameters}
\usage{
get_aed_sed_const2d_param(aeme, path, lake_dir = NULL)
}
\arguments{
\item{aeme}{aeme; object.}

\item{path}{filepath; where input files are located relative to the current 
working directory.}
}
\value{
Data frame with AED sed_const2d parameters
}
\description{
Get AED sed_const2d parameters
}
