% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aeme_params.R
\name{get_aeme_parameters}
\alias{get_aeme_parameters}
\title{Make parameters dataframe for AEME}
\usage{
get_aeme_parameters(model, file, module, name, par)
}
\arguments{
\item{model}{vector; of models to be used. Can be \code{dy_cd}, \code{glm_aed},
\code{gotm_wet}.}

\item{file}{character; vector of file names to retrieve.}

\item{module}{character; vector of modules to retrieve parameters for.}

\item{name}{character; vector of parameter names to retrieve.}

\item{par}{character; vector of parameters to retrieve.}
}
\value{
dataframe; of parameters filtered by model, module and parameter.
}
\description{
Make parameters dataframe for AEME
}
