% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aeme_params.R
\name{get_aeme_parameters}
\alias{get_aeme_parameters}
\title{Make parameters dataframe for AEME}
\usage{
get_aeme_parameters(model, module, name, par)
}
\arguments{
\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{module}{character; vector of modules to retrieve parameters for.}

\item{par}{character; vector of parameters to retrieve.}
}
\value{
dataframe; of parameters filtered by model, module and parameter.
}
\description{
Make parameters dataframe for AEME
}
