% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aeme_path.R
\name{get_aeme_path}
\alias{get_aeme_path}
\title{Get the path to the AEME data}
\usage{
get_aeme_path(aeme)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
Path to the AEME data
}
\description{
Get the path to the AEME data
}
