% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_index.R
\name{get_date_index}
\alias{get_date_index}
\title{Get date index for each model in the AEME object}
\usage{
get_date_index(aeme, model, remove_spin_up = TRUE)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}
}
\value{
A list with date index for each model
}
\description{
Get date index for each model in the AEME object
}
