% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_index.R
\name{get_date_index}
\alias{get_date_index}
\title{Get date index for each model in the AEME object}
\usage{
get_date_index(aeme, model, remove_spin_up = TRUE)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}
}
\value{
A list with date index for each model
}
\description{
Get date index for each model in the AEME object
}
