% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glm_sed_params.R
\name{get_glm_sed_params}
\alias{get_glm_sed_params}
\title{Get number of sediment zones in GLM-AED model}
\usage{
get_glm_sed_params(aeme, path, lake_dir = NULL)
}
\arguments{
\item{aeme}{Aeme object.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}

\item{lake_dir}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{path} instead of
\code{lake_dir}}
}
\value{
Number of sediment zones
}
\description{
Get number of sediment zones in GLM-AED model
}
