% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glm_sed_zones.R
\name{get_glm_sed_zones}
\alias{get_glm_sed_zones}
\title{Get number of sediment zones in GLM-AED model}
\usage{
get_glm_sed_zones(aeme, path, lake_dir = NULL)
}
\arguments{
\item{aeme}{aeme; object.}

\item{path}{filepath; where input files are located relative to the current 
working directory.}
}
\value{
Number of sediment zones
}
\description{
Get number of sediment zones in GLM-AED model
}
