% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inflows.R
\name{get_inflows}
\alias{get_inflows}
\title{Get inflow data from Aeme object}
\usage{
get_inflows(aeme, return_df = FALSE)
}
\arguments{
\item{aeme}{Aeme object.}

\item{return_df}{Logical. If TRUE, returns a single data frame with all
inflows. If FALSE, returns a list of data frames for each inflow. Default is
FALSE.}
}
\value{
A list of data frames or a single data frame with all inflows
combined.
}
\description{
Get inflow data from Aeme object
}
