% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inflows.R
\name{get_inflows}
\alias{get_inflows}
\title{Get inflow data from Aeme object}
\usage{
get_inflows(aeme, return_df = FALSE)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
A list of data frames or a single data frame with all inflows
combined.
}
\description{
Get inflow data from Aeme object
}
