% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake_dir.R
\name{get_lake_dir}
\alias{get_lake_dir}
\title{Get the directory of the lake model setup}
\usage{
get_lake_dir(aeme, path)
}
\arguments{
\item{aeme}{aeme; object.}

\item{path}{filepath; where input files are located relative to the current
working directory.}
}
\value{
character; the directory of the lake model setup
}
\description{
Get the directory of the lake model setup
}
