% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mean_sea_level_pressure}
\alias{get_mean_sea_level_pressure}
\title{Return mean sea level pressure given air temperature, elevation and station pressure.}
\usage{
get_mean_sea_level_pressure(prsttn, elevation, tmpair)
}
\arguments{
\item{elevation}{A numeric vector of elevation in m}

\item{MET_prsttn}{A numeric vector of observed station pressure in Pa}

\item{MET_tmpair}{A numeric vector of air temperature in degC}
}
\value{
A numeric vector of mean sea level pressure in Pa
}
\description{
Return mean sea level pressure given air temperature, elevation and station pressure.
}
\note{
The standard procedure for the US is to use for MET_tmpair the average
of the current station temperature and the station temperature from 12 hours ago.
}
\examples{
get_mean_sea_level_pressure(101226.5, 105:205, 17.19)

}
\references{
Hess SL, Introduction to theoretical meteorology, Holt Rinehart and Winston, NY 1959,
ch. 6.5; Stull RB, Meteorology for scientists and engineers, 2nd edition,
Brooks/Cole 2000, ch. 1.
}
