% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod_obs_vars.R
\name{get_mod_obs_vars}
\alias{get_mod_obs_vars}
\title{Get modeled observation variables}
\usage{
get_mod_obs_vars(aeme, model)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}
}
\value{
Data frame with modeled observation variables and summary
statistics.
}
\description{
Get modeled observation variables
}
