% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod_obs_vars.R
\name{get_mod_obs_vars}
\alias{get_mod_obs_vars}
\title{Get modeled observation variables}
\usage{
get_mod_obs_vars(aeme, model)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}
}
\value{
Data frame with modeled observation variables and summary 
statistics.
}
\description{
Get modeled observation variables
}
