% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_config_files.R
\name{get_model_config_files}
\alias{get_model_config_files}
\title{Get model configuration files paths}
\usage{
get_model_config_files(aeme, model, path, lake_dir = NULL)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}, \code{"simstrat_aed2"}. Defaults to all models if
not found in \code{aeme}.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}

\item{lake_dir}{Path to the lake AEME directory. If NULL, it will be
computed from \code{aeme} and \code{path}.}
}
\value{
A list with model configuration files paths
}
\description{
Get model configuration files paths
}
