% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_config_files.R
\name{get_model_config_files}
\alias{get_model_config_files}
\title{Get model configuration files paths}
\usage{
get_model_config_files(aeme, model, path, lake_dir = NULL)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be \code{dy_cd}, \code{glm_aed},
\code{gotm_wet}.}

\item{path}{filepath; where input files are located relative to the current
working directory.}

\item{lake_dir}{Path to the lake AEME directory. If NULL, it will be
computed from \code{aeme} and \code{path}.}
}
\value{
A list with model configuration files paths
}
\description{
Get model configuration files paths
}
