% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_config_files.R
\name{get_model_config_files}
\alias{get_model_config_files}
\title{Get model configuration files paths}
\usage{
get_model_config_files(aeme = NULL, model, path = NULL, lake_dir)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}, \code{"simstrat_aed2"}. Defaults to all models if
not found in \code{aeme}.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}

\item{lake_dir}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{path} instead of
\code{lake_dir}}
}
\value{
A list with model configuration files paths
}
\description{
Get model configuration files paths
}
