% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_config_files.R
\name{get_model_config_files}
\alias{get_model_config_files}
\title{Get model configuration files paths}
\usage{
get_model_config_files(aeme, model, path)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{path}{filepath; where input files are located relative to `config`.}
}
\value{
A named vector with paths to model configuration files
}
\description{
Get model configuration files paths
}
