% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme.R
\name{get_model_version}
\alias{get_model_version}
\title{Get model version}
\usage{
get_model_version(model)
}
\arguments{
\item{model}{model name. Only "glm_aed", "gotm_wet", and "dy_cd" are
supported.}
}
\value{
version string (invisible)
}
\description{
Get model version
}
