% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs.R
\name{get_obs}
\alias{get_obs}
\title{Get observations for a given variable from an Aeme object}
\usage{
get_obs(aeme, var_sim, depth_range = NULL, time_filter = FALSE)
}
\arguments{
\item{aeme}{Aeme object.}

\item{var_sim}{character; variable in the AEME format (e.g. "HYD_temp"). Can
be a vector. If missing, all variables are returned.}

\item{depth_range}{numeric vector of length 2; depth range (in meters) to
filter observations. If NULL, all depths are returned.}

\item{time_filter}{logical; if TRUE, filter observations to the time range of
the Aeme object. If FALSE, all observations are returned regardless of time.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{Date}: Date of observation
\item \code{var_aeme}: Name of the variable in the AEME format
\item \code{depth_from}: Depth from which the variable is extracted
\item \code{depth_to}: Depth to which the variable is extracted
\item \code{value}: Value of the variable
}
}
\description{
Get observations for a given variable from an Aeme object
}
