% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{get_obs_column_names}
\alias{get_obs_column_names}
\title{Get column names for the observational data frame}
\usage{
get_obs_column_names()
}
\value{
vector of column names for the observational data frame
}
\description{
Get column names for the observational data frame
}
