% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_var}
\alias{get_var}
\title{Get variable from aeme}
\usage{
get_var(
  aeme,
  model,
  var_sim,
  depth = NULL,
  depth_ref = c("surface", "bottom"),
  return_df = TRUE,
  ens_n = 1,
  use_obs = FALSE,
  remove_spin_up = TRUE,
  cumulative = FALSE
)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{var_sim}{character; variable in the AEME format (e.g. "HYD_temp").}

\item{depth}{numeric; depth of the variable to extract. Default is NULL. If
NULL, the variable profiles are extracted.}

\item{depth_ref}{character; reference depth for extracting variable profiles.
Options are "surface" (default) or "bottom".}

\item{return_df}{logical; if TRUE, return a dataframe; if FALSE, return a
list. Default is TRUE.}

\item{ens_n}{integer; ensemble number to plot. Default is 1.}

\item{use_obs}{logical; if TRUE, use observations to extract the variable at
time and depth of observations. Default is FALSE. Use this option if you
want to compare model output to observations.}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}

\item{cumulative}{logical; if TRUE, return cumulative sum of variable}
}
\value{
dataframe or list
}
\description{
Get variable from aeme
}
