% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_var}
\alias{get_var}
\title{Get variable from aeme_data}
\usage{
get_var(aeme_data, model, var_sim, return_df = TRUE, cumulative = FALSE)
}
\arguments{
\item{aeme_data}{aeme; data object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{var_sim}{string; of variable to plot}

\item{return_df}{logical; if TRUE, return a dataframe; if FALSE, return a
list. Default is TRUE.}

\item{cumulative}{logical; if TRUE, return cumulative sum of variable}
}
\value{
dataframe or list
}
\description{
Get variable from aeme_data
}
