% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_var}
\alias{get_var}
\title{Get variable from aeme}
\usage{
get_var(
  aeme,
  model,
  var_sim,
  return_df = TRUE,
  use_obs = FALSE,
  cumulative = FALSE
)
}
\arguments{
\item{aeme}{aeme; data object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{var_sim}{string; of variable to plot}

\item{return_df}{logical; if TRUE, return a dataframe; if FALSE, return a
list. Default is TRUE.}

\item{use_obs}{logical; if TRUE, use observations to extract the variable at
time and depth of observations. Default is FALSE. Use this option if you
want to compare model output to observations.}

\item{cumulative}{logical; if TRUE, return cumulative sum of variable}
}
\value{
dataframe or list
}
\description{
Get variable from aeme
}
