% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_indices.R
\name{get_var_indices}
\alias{get_var_indices}
\title{Get variable indices}
\usage{
get_var_indices(
  nc = NULL,
  model,
  aeme,
  path,
  vars_sim,
  month = NULL,
  depth_range = NULL,
  use_obs = TRUE
)
}
\arguments{
\item{nc}{An object of class \code{ncdf4} (as returned by either function
 \code{\link[ncdf4]{nc_open}} or function \code{\link[ncdf4]{nc_create}}.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}, \code{"simstrat_aed2"}. Defaults to all models if
not found in \code{aeme}.}

\item{aeme}{Aeme object.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}

\item{vars_sim}{character; vector of AEME variable names to get indices for.}

\item{month}{numeric; vector of months to subset the data.}

\item{depth_range}{numeric; vector of depth ranges, length two to subset the
data.}

\item{use_obs}{logical; if TRUE, use the observation months and depth ranges
from the AEME object.}
}
\value{
list; of variable indices. Each list element corresponds to a
variable in vars_sim and contains a list with time indices, depth values, and
dates. Time indices correspond to the positions in the model output time
series that match the Date but are the corresponding index in the model
output.
}
\description{
Get variable indices for an AEME variable.
}
