% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_indices.R
\name{get_var_indices}
\alias{get_var_indices}
\title{Get variable indices}
\usage{
get_var_indices(
  nc = NULL,
  model,
  aeme,
  path,
  vars_sim,
  month = NULL,
  depth_range = NULL
)
}
\arguments{
\item{nc}{An object of class \code{ncdf4} (as returned by either function
 \code{\link[ncdf4]{nc_open}} or function \code{\link[ncdf4]{nc_create}}.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{aeme}{aeme; object.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{month}{numeric; vector of months to subset the data.}

\item{depth_range}{numeric; vector of depth ranges, length two to subset the
data.}
}
\value{
list; of variable indices.
}
\description{
Get variable indices for an AEME variable.
}
