% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars_sim.R
\name{get_vars_sim}
\alias{get_vars_sim}
\title{Get all variables to be simulated}
\usage{
get_vars_sim(aeme, model_controls)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}
}
\value{
vector of variable names
}
\description{
Get all variables to be simulated
}
\examples{
data("model_controls", package = "AEME")
get_vars_sim(model_controls)
}
