% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wbal_components.R
\name{get_wbal_components}
\alias{get_wbal_components}
\title{Get water balance components from AEME object}
\usage{
get_wbal_components(aeme, model, remove_spin_up = TRUE, cumulative = FALSE)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}

\item{cumulative}{logical; if TRUE, return cumulative sum of variable}
}
\value{
List with observed lake levels, AEME water balance, and model components
}
\description{
Get water balance components from AEME object
}
