% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_glm_aed.R
\name{glm_exe_path}
\alias{glm_exe_path}
\title{Locate an installed GLM executable}
\usage{
glm_exe_path(version = getOption("AEME.glm_version", NULL), os = NULL)
}
\arguments{
\item{version}{Character. The GLM version to locate, e.g. \code{"3.9.108"}.}

\item{os}{Character. One of \code{"windows"}, \code{"macos"}, or \code{"linux"}.
Defaults to the current platform.}
}
\value{
Character. The file path to the \code{glm}/\code{glm.exe} executable.
}
\description{
Returns the path to a GLM executable previously installed with
\code{\link[=install_glm_aed]{install_glm_aed()}}. Errors if that version isn't installed for this
platform yet.
}
\examples{
\dontrun{
install_glm_aed(version = "3.9.108")
glm_exe_path(version = "3.9.108")
}

}
