% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{inflows,aeme-method}
\alias{inflows,aeme-method}
\title{Access inflows slot}
\usage{
\S4method{inflows}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of inflows characteristics
}
\description{
Access inflows slot
}
