% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{inflows<-,Aeme-method}
\alias{inflows<-,Aeme-method}
\title{Set inflows in Aeme object}
\usage{
\S4method{inflows}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New inflows data to be assigned.}
}
\value{
A modified Aeme object with updated inflows slot.
}
\description{
Update the inflows slot of an Aeme object
}
\details{
This method updates the "inflows" slot of An Aeme object with new data.
}
