% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{inflows<-}
\alias{inflows<-}
\title{Set inflows in aeme object}
\usage{
inflows(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New inflows data to be assigned.}
}
\value{
A modified aeme object with updated inflows slot.
}
\description{
Update the inflows slot of an aeme object
}
