% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{inflows<-}
\alias{inflows<-}
\title{Set inflows in Aeme object}
\usage{
inflows(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New inflows data to be assigned.}
}
\value{
A modified Aeme object with updated inflows slot.
}
\description{
Update the inflows slot of an Aeme object
}
