% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{inflows}
\alias{inflows}
\alias{inflows,Aeme-method}
\title{Access the inflows slot of an Aeme object}
\usage{
inflows(aeme)

\S4method{inflows}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
List of inflows characteristics.
}
\description{
Access the inflows slot of an Aeme object
}
