% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{input,aeme-method}
\alias{input,aeme-method}
\title{Access input slot}
\usage{
\S4method{input}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of input characteristics
}
\description{
Access input slot
}
