% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{input<-,Aeme-method}
\alias{input<-,Aeme-method}
\title{Set input in Aeme object}
\usage{
\S4method{input}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New input data to be assigned.}
}
\value{
A modified Aeme object with updated input slot.
}
\description{
Update the input slot of an Aeme object
}
\details{
This method updates the "input" slot of a Aeme object with new data.
}
