% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{input<-}
\alias{input<-}
\alias{input<-,Aeme-method}
\title{Set the input slot of an Aeme object}
\usage{
input(aeme) <- value

\S4method{input}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New input list to assign.}
}
\value{
A modified Aeme object with updated input slot.
}
\description{
Set the input slot of an Aeme object
}
