% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{input<-}
\alias{input<-}
\title{Set input in aeme object}
\usage{
input(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New input data to be assigned.}
}
\value{
A modified aeme object with updated input slot.
}
\description{
Update the input slot of an aeme object
}
