% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{input<-}
\alias{input<-}
\title{Set input in Aeme object}
\usage{
input(aeme) <- value
}
\arguments{
\item{aeme}{A Aeme object.}

\item{value}{New input data to be assigned.}
}
\value{
A modified Aeme object with updated input slot.
}
\description{
Update the input slot of an Aeme object
}
