% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_model_parameters.R
\name{input_model_parameters}
\alias{input_model_parameters}
\title{Input model parameters}
\usage{
input_model_parameters(aeme, model, param, path)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{param}{data.frame; parameters to input into the model
configuration files}

\item{path}{filepath; where input files are located relative to the current 
working directory.}
}
\value{
Aeme object with parameters input into model configuration files
}
\description{
Input model parameters
}
