% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_model_parameters.R
\name{input_model_parameters}
\alias{input_model_parameters}
\title{Input model parameters}
\usage{
input_model_parameters(aeme, model, param, path)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{param}{data.frame; parameters to input into the model
configuration files}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}
}
\value{
Aeme object with parameters input into model configuration files
}
\description{
Input model parameters
}
