% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_strat.R
\name{is_strat}
\alias{is_strat}
\title{Check if water temperature profile is stratified}
\usage{
is_strat(wtr, depths, t_diff = 1)
}
\arguments{
\item{wtr}{numeric vector; water temperature profile}

\item{depths}{numeric vector; depths corresponding to water temperature 
profile}

\item{t_diff}{numeric; minimum temperature difference between surface and
bottom to consider the profile stratified. Default is 1 degree Celsius.}
}
\value{
logical; TRUE if profile is stratified, FALSE otherwise
}
\description{
Check if water temperature profile is stratified
}
