% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_naming}
\alias{key_naming}
\title{Model variable naming}
\format{
\subsection{\code{key_naming}}{

A data frame with 130 rows and 18 columns:
\describe{
\item{var_aeme}{AEME variable name}
\item{dy_cd}{DYRESM variable name}
\item{glm_aed}{GLM variable name}
\item{glm_aed2}{GLM-AED2 variable name}
\item{simstrat_aed2}{Simstrat-AED2 variable name}
\item{gotm_wet}{GOTM variable name}
\item{gotm_fabm}{GOTM-FABM variable name containing key-value names}
\item{name_text}{Regular text string variable name}
\item{name_full}{String variable name}
\item{name_parse}{Variable name for parsing}
\item{units}{Units of the variable. As defined using the
\href{https://cran.r-project.org/package=units}{\pkg{units}} package.}
\item{conversion_aed}{Unit conversion for GLM-AED}
\item{default}{Default value for the variable}
\item{derived}{Logical value if the variable is derived}
\item{derived_from}{Variable name that the variable is derived from}
\item{min}{Minimum value for the variable}
\item{max}{Maximum value for the variable}
\item{keywords}{Keywords associated with the variable}
}
}
}
\source{
Package development.
}
\usage{
key_naming
}
\description{
A reference table for variable names between the models in AEME.
}
\keyword{datasets}
