% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{lake,Aeme-method}
\alias{lake,Aeme-method}
\title{Access lake slot}
\usage{
\S4method{lake}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
list of lake characteristics
}
\description{
Access lake slot
}
