% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{lake,aeme-method}
\alias{lake,aeme-method}
\title{Access lake slot}
\usage{
\S4method{lake}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of lake characteristics
}
\description{
Access lake slot
}
