% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{lake<-,Aeme-method}
\alias{lake<-,Aeme-method}
\title{Set lake in Aeme object}
\usage{
\S4method{lake}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New lake data to be assigned.}
}
\value{
A modified Aeme object with updated lake slot.
}
\description{
Update the lake slot of an Aeme object
}
\details{
This method updates the "lake" slot of An Aeme object with new data.
}
