% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{lake<-}
\alias{lake<-}
\alias{lake<-,Aeme-method}
\title{Set the lake slot of an Aeme object}
\usage{
lake(aeme) <- value

\S4method{lake}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New lake list to assign.}
}
\value{
A modified Aeme object with updated lake slot.
}
\description{
Set the lake slot of an Aeme object
}
