% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{lake<-}
\alias{lake<-}
\title{Set lake in Aeme object}
\usage{
lake(aeme) <- value
}
\arguments{
\item{aeme}{A Aeme object.}

\item{value}{New lake data to be assigned.}
}
\value{
A modified Aeme object with updated lake slot.
}
\description{
Update the lake slot of an Aeme object
}
