% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{lake<-}
\alias{lake<-}
\title{Set lake in aeme object}
\usage{
lake(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New lake data to be assigned.}
}
\value{
A modified aeme object with updated lake slot.
}
\description{
Update the lake slot of an aeme object
}
