% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_obs_vars.R
\name{list_obs_vars}
\alias{list_obs_vars}
\title{Get observation variable names}
\usage{
list_obs_vars(aeme)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
Observation variable names vector
}
\description{
Get observation variable names
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
list_obs_vars(aeme)
}
