% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_configuration.R
\name{load_configuration}
\alias{load_configuration}
\title{Load model configuration to the aeme object}
\usage{
load_configuration(aeme, model, model_controls = NULL, use_bgc = FALSE, path)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}

\item{use_bgc}{logical; switch to use the biogeochemical model.}

\item{path}{filepath; where input files are located relative to the current 
working directory.}
}
\value{
Updated aeme object with model configuration
}
\description{
Load model configuration to the aeme object
}
