% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_configuration.R
\name{load_configuration}
\alias{load_configuration}
\title{Load model configuration to the aeme object}
\usage{
load_configuration(model, aeme_data, path, use_bgc = FALSE)
}
\arguments{
\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{aeme_data}{aeme; data object.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{use_bgc}{logical; switch to use the biogeochemical model.}
}
\value{
Updated aeme object with model configuration
}
\description{
Load model configuration to the aeme object
}
