% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_output}
\alias{load_output}
\title{Load AEME output to the aeme object}
\usage{
load_output(
  aeme,
  model,
  path = NULL,
  lake_dir = NULL,
  model_controls,
  parallel = FALSE,
  cl = NULL,
  ens_n = 1
)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}

\item{model_controls}{data.frame; model configuration, typically loaded via
\code{\link[=get_model_controls]{get_model_controls()}}.}

\item{parallel}{logical; run models in parallel. Defaults to FALSE.}

\item{cl}{an object of class \code{"cluster"}.}

\item{ens_n}{numeric; ensemble number to allocate to model output which is
loaded. Defaults to 1.}
}
\value{
Updated aeme object with model output
}
\description{
Load AEME output to the aeme object
}
