% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_output}
\alias{load_output}
\title{Load AEME output to the aeme object}
\usage{
load_output(
  aeme,
  model,
  path = NULL,
  lake_dir = NULL,
  model_controls,
  parallel = FALSE,
  cl = NULL,
  ens_n = 1
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be \code{dy_cd}, \code{glm_aed},
\code{gotm_wet}.}

\item{path}{filepath; where input files are located relative to the current
working directory.}

\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}

\item{parallel}{logical; run models in parallel. Defaults to FALSE.}

\item{cl}{an object of class \code{"cluster"}.}

\item{ens_n}{numeric; ensemble number to allocate to model output which is
loaded. Defaults to 1.}
}
\value{
Updated aeme object with model output
}
\description{
Load AEME output to the aeme object
}
