% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_output}
\alias{load_output}
\title{Load AEME output to the aeme object}
\usage{
load_output(model, aeme, path, model_controls, parallel = FALSE, nlev = NULL)
}
\arguments{
\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{aeme}{aeme; data object.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}

\item{parallel}{logical; run models in parallel. Defaults to FALSE.}

\item{nlev}{numeric; number of levels to return in model output. If NULL,
calculates number of levels based on lake depth as provided within the
hypsograph. If the lake is less thatn 3m deep, it outputs every 0.1m, else
if it is greater than 3m, it outputs every 0.33m. Defaults to NULL.}
}
\value{
Updated aeme object with model output
}
\description{
Load AEME output to the aeme object
}
