% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_controls}
\alias{model_controls}
\title{Model controls}
\format{
## `model_controls`
A data frame with 65 rows and 6 columns:
\describe{
  \item{var_aeme}{AEME variable}
  \item{simulate}{Logical value to simulate the variable}
  \item{inf_default}{Default value in inflows.}
  \item{initial_wc}{Default value for initialising in the water column.}
  \item{initial_sed}{Default value for initialising in the sediments.}
  \item{conversion_aed}{Unit conversion for GLM-AED.}
}
}
\source{
Package development.
}
\usage{
model_controls
}
\description{
A reference table for catchment nutrient and sediment yields for calculating
inflow nutrients in AEME.
}
\author{
Tadhg Moore, Chris McBride
}
\keyword{datasets}
