% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{names,Aeme-method}
\alias{names,Aeme-method}
\title{Names method for Aeme objects}
\usage{
\S4method{names}{Aeme}(x)
}
\arguments{
\item{x}{An Aeme object.}
}
\value{
Character vector of slot names.
}
\description{
Get names of an Aeme object
}
\details{
Returns the names of all slots in the Aeme object.
}
