% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{names,Aeme-method}
\alias{names,Aeme-method}
\title{Update names Method}
\usage{
\S4method{names}{Aeme}(x)
}
\arguments{
\item{x}{An Aeme object.}
}
\value{
vector of names of the slots in the Aeme object to the console.
}
\description{
Get names of an Aeme object
}
\details{
This method prints the names of the slots in the Aeme object.
}
