% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{names,aeme-method}
\alias{names,aeme-method}
\title{Update names Method}
\usage{
\S4method{names}{aeme}(x)
}
\arguments{
\item{x}{An aeme object.}
}
\value{
vector of names of the slots in the aeme object to the console.
}
\description{
Plot an aeme object
}
\details{
This method prints the names of the slots in the aeme object.
}
