% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_do.R
\name{o2_at_sat}
\alias{o2_at_sat}
\title{Estimate oxygen saturation concentration}
\usage{
o2_at_sat(
  temp,
  depth,
  baro,
  altitude = 0,
  salinity = rep(0, length(temp)),
  model = "garcia_benson"
)
}
\arguments{
\item{temp}{water temperature (°C)}

\item{baro}{barometric pressure (mb)}

\item{altitude}{altitude (m). Only used if `baro` is missing.}

\item{salinity}{salinity (ppt).}

\item{model}{character, solubility model to use. Options are
"garcia", "garcia_benson", "weiss", or "benson".}
}
\value{
numeric vector of oxygen saturation concentrations (mg/L)
}
\description{
Estimate oxygen saturation concentration
}
\references{
Colt, John. \emph{1 - Solubility of Atmospheric Gases in Freshwater.} In
Computation of Dissolved Gas Concentration in Water as Functions of
Temperature, Salinity and Pressure (Second Edition), edited by John Colt,
1-71. London: Elsevier, 2012.
http://www.sciencedirect.com/science/article/pii/B9780124159167000012.

Garcia, H., and L. Gordon (1992), \emph{Oxygen solubility in seawater: Better
fitting equations}, Limnol. Oceanogr., 37(6).

Benson, B. B. & Krause, D. (1984). \emph{The concentration and isotopic
fractionation of oxygen dissolved in freshwater and seawater in equilibrium
with the atmosphere.} Limnology and Oceanography, 29(3), 620-632.
doi:10.4319/lo.1984.29.3.0620

Staehr, Peter A., Darren Bade, Matthew C. Van de Bogert, Gregory R. Koch,
Craig Williamson, Paul Hanson, Jonathan J. Cole, and Tim Kratz. \emph{Lake
Metabolism and the Diel Oxygen Technique: State of the Science.} Limnology
and Oceanography: Methods 8, no. 11 (November 1, 2010): 628-44.
doi:10.4319/lom.2010.8.0628

USGS. \emph{New Tables of Dissolved Oxygen Saturation Values.} Quality of
Water Branch, 1981. http://water.usgs.gov/admin/memo/QW/qw81.11.html.

USGS. \emph{New Tables of Dissolved Oxygen Saturation Values; Amendment of
Quality of Water Technical Memorandum No. 81.11.} Quality of Water Branch,
1981. http://water.usgs.gov/admin/memo/QW/qw81.15.html.

USGS. \emph{Change to Solubility Equations for Oxygen in Water.} Technical
Memorandum 2011.03. USGS Office of Water Quality, 2011.

Weiss, R. (1970). \emph{The solubility of nitrogen, oxygen and argon in water
and seawater}. Deep Sea Research and Oceanographic Abstracts, 17(4), 721-735.
doi:10.1016/0011-7471(70)90037-9
}
