% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{observations,Aeme-method}
\alias{observations,Aeme-method}
\title{Access observations slot}
\usage{
\S4method{observations}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
list of observations characteristics
}
\description{
Access observations slot
}
