% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{observations<-,Aeme-method}
\alias{observations<-,Aeme-method}
\title{Set observations in Aeme object}
\usage{
\S4method{observations}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New observations data to be assigned.}
}
\value{
A modified Aeme object with updated observations slot.
}
\description{
Update the observations slot of an Aeme object
}
\details{
This method updates the "observations" slot of An Aeme object with new data.
}
