% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{observations<-}
\alias{observations<-}
\title{Set observations in Aeme object}
\usage{
observations(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New observations data to be assigned.}
}
\value{
A modified Aeme object with updated observations slot.
}
\description{
Update the observations slot of an Aeme object
}
