% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{observations<-}
\alias{observations<-}
\title{Set observations in aeme object}
\usage{
observations(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New observations data to be assigned.}
}
\value{
A modified aeme object with updated observations slot.
}
\description{
Update the observations slot of an aeme object
}
