% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{observations}
\alias{observations}
\title{Access observations slot}
\usage{
observations(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
list of observations characteristics
}
\description{
Access observations slot
}
