% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{outflows,Aeme-method}
\alias{outflows,Aeme-method}
\title{Access outflows slot}
\usage{
\S4method{outflows}{Aeme}(aeme)
}
\arguments{
\item{aeme}{A Aeme object.}
}
\value{
list of outflows characteristics
}
\description{
Access outflows slot
}
