% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{outflows<-,Aeme-method}
\alias{outflows<-,Aeme-method}
\title{Set outflows in Aeme object}
\usage{
\S4method{outflows}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New outflows data to be assigned.}
}
\value{
A modified Aeme object with updated outflows slot.
}
\description{
Update the outflows slot of an Aeme object
}
\details{
This method updates the "outflows" slot of An Aeme object with new data.
}
