% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{outflows<-}
\alias{outflows<-}
\title{Set outflows in aeme object}
\usage{
outflows(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New outflows data to be assigned.}
}
\value{
A modified aeme object with updated outflows slot.
}
\description{
Update the outflows slot of an aeme object
}
