% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{outflows<-}
\alias{outflows<-}
\title{Set outflows in Aeme object}
\usage{
outflows(aeme) <- value
}
\arguments{
\item{aeme}{A Aeme object.}

\item{value}{New outflows data to be assigned.}
}
\value{
A modified Aeme object with updated outflows slot.
}
\description{
Update the outflows slot of an Aeme object
}
