% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{outflows}
\alias{outflows}
\alias{outflows,Aeme-method}
\title{Access the outflows slot of an Aeme object}
\usage{
outflows(aeme)

\S4method{outflows}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
List of outflows characteristics.
}
\description{
Access the outflows slot of an Aeme object
}
