% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{output,aeme-method}
\alias{output,aeme-method}
\title{Access output slot}
\usage{
\S4method{output}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of output characteristics
}
\description{
Access output slot
}
