% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{output<-,aeme-method}
\alias{output<-,aeme-method}
\title{Set output in aeme object}
\usage{
\S4method{output}{aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An aeme object.}

\item{value}{New output data to be assigned.}
}
\value{
A modified aeme object with updated output slot.
}
\description{
Update the output slot of an aeme object
}
\details{
This method updates the "output" slot of a aeme object with new data.
}
