% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{output<-}
\alias{output<-}
\title{Set output in aeme object}
\usage{
output(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New output data to be assigned.}
}
\value{
A modified aeme object with updated output slot.
}
\description{
Update the output slot of an aeme object
}
