% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{output<-}
\alias{output<-}
\title{Set output in Aeme object}
\usage{
output(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New output data to be assigned.}
}
\value{
A modified Aeme object with updated output slot.
}
\description{
Update the output slot of an Aeme object
}
