% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{parameters,Aeme-method}
\alias{parameters,Aeme-method}
\title{Access parameters slot}
\usage{
\S4method{parameters}{Aeme}(aeme)
}
\arguments{
\item{aeme}{A Aeme object.}
}
\value{
dataframe of parameters
}
\description{
Access parameters slot
}
