% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{parameters,aeme-method}
\alias{parameters,aeme-method}
\title{Access parameters slot}
\usage{
\S4method{parameters}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
dataframe of parameters
}
\description{
Access parameters slot
}
