% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{parameters<-,Aeme-method}
\alias{parameters<-,Aeme-method}
\title{Set parameters in Aeme object}
\usage{
\S4method{parameters}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New parameters data to be assigned.}
}
\value{
A modified Aeme object with updated parameters slot.
}
\description{
Update the parameters slot of an Aeme object
}
\details{
This method updates the "parameters" slot of An Aeme object with new data.
}
