% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{parameters<-}
\alias{parameters<-}
\alias{parameters<-,Aeme-method}
\title{Set the parameters slot of an Aeme object}
\usage{
parameters(aeme) <- value

\S4method{parameters}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New parameters data.frame to assign.}
}
\value{
A modified Aeme object with updated parameters slot.
}
\description{
Set the parameters slot of an Aeme object
}
