% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{parameters<-}
\alias{parameters<-}
\title{Set parameters in Aeme object}
\usage{
parameters(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New parameters data to be assigned.}
}
\value{
A modified Aeme object with updated parameters slot.
}
\description{
Update the parameters slot of an Aeme object
}
