% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{parameters<-}
\alias{parameters<-}
\title{Set parameters in aeme object}
\usage{
parameters(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New parameters data to be assigned.}
}
\value{
A modified aeme object with updated parameters slot.
}
\description{
Update the parameters slot of an aeme object
}
