% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{plot,aeme-method}
\alias{plot,aeme-method}
\title{Update summary Method}
\usage{
\S4method{plot}{aeme}(x, y, ..., add = FALSE)
}
\arguments{
\item{x}{An aeme object.}

\item{y}{An aeme slot (optional). Defaults to "output".}

\item{...}{additional arguments affecting the plot produced.}

\item{add}{logical; add to current plot?}
}
\value{
prints the aeme object to the console.
}
\description{
Plot an aeme object
}
\details{
This method plots the aeme object.
}
