% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{plot,Aeme-method}
\alias{plot,Aeme-method}
\title{Update summary Method}
\usage{
\S4method{plot}{Aeme}(x, y, ..., add = FALSE)
}
\arguments{
\item{x}{An Aeme object.}

\item{y}{An Aeme slot (optional). Defaults to "output".}

\item{...}{additional arguments affecting the plot produced.}

\item{add}{logical; add to current plot?}
}
\value{
prints the Aeme object to the console.
}
\description{
Plot an Aeme object
}
\details{
This method plots the Aeme object.
}
