% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{plot,aeme-method}
\alias{plot,aeme-method}
\title{Update summary Method}
\usage{
\S4method{plot}{aeme}(x, y, ...)
}
\arguments{
\item{x}{An aeme object.}

\item{y}{An aeme slot (optional).}

\item{...}{additional arguments affecting the plot produced.}
}
\value{
prints the aeme object to the console.
}
\description{
Plot an aeme object
}
\details{
This method plots the aeme object.
}
