% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fluxes.R
\name{plot_fluxes}
\alias{plot_fluxes}
\title{Plot fluxes}
\usage{
plot_fluxes(aeme, model, cumulative = FALSE)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{cumulative}{logical; if \code{TRUE}, plot cumulative fluxes. If
\code{FALSE}, plot instantaneous fluxes.}
}
\value{
ggplot2 object
}
\description{
Plot heat fluxes from AEME simulations. This includes incoming
shortwave radiation, net longwave radiation, evaporative heat flux, and
sensible heat flux.
}
