% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fluxes.R
\name{plot_fluxes}
\alias{plot_fluxes}
\title{Plot fluxes}
\usage{
plot_fluxes(aeme, model, cumulative = FALSE)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{cumulative}{logical; if \code{TRUE}, plot cumulative fluxes. If
\code{FALSE}, plot instantaneous fluxes.}
}
\value{
ggplot2 object
}
\description{
Plot heat fluxes from AEME simulations. This includes incoming
shortwave radiation, net longwave radiation, evaporative heat flux, and
sensible heat flux.
}
