% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_met_tile.R
\name{plot_met_tile}
\alias{plot_met_tile}
\title{Plot a tile plot of meteorological data}
\usage{
plot_met_tile(aeme, var_inp = "MET_tmpair", use_hydro_year = TRUE)
}
\arguments{
\item{aeme}{Aeme object.}

\item{var_inp}{Character. Variable to plot. Can be one of:
\itemize{
\item \code{"MET_tmpair"}: Air temperature
\item \code{"MET_pprain"}: Rainfall
\item \code{"MET_wndspd"}: Wind speed
\item \code{"MET_humrel"}: Relative humidity
\item \code{"MET_radswd"}: Shortwave radiation
\item \code{"MET_radlwd"}: Longwave radiation
\item \code{"MET_pres"}: Atmospheric pressure
\item \code{"MET_ppsnow"}: Snowfall
\item \code{"MET_wnddir"}: Wind direction
}}

\item{use_hydro_year}{Logical. If \code{TRUE}, the hydrological year is used.}
}
\value{
A ggplot object
}
\description{
Plot a tile plot of meteorological data
}
