% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs.R
\name{plot_obs}
\alias{plot_obs}
\title{Plot observations}
\usage{
plot_obs(aeme, var_sim = "HYD_temp", add_line = FALSE)
}
\arguments{
\item{aeme}{aeme; object.}

\item{var_sim}{string; of variable to plot}

\item{add_line}{logical, add line to the plot}
}
\value{
ggplot object
}
\description{
Plot observations
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
plot_obs(aeme = aeme, var_sim = c("HYD_temp", "LKE_lvlwtr"))
}
