% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{plot_output}
\alias{plot_output}
\title{Plot AEME output}
\usage{
plot_output(
  aeme,
  model,
  var_sim = "HYD_temp",
  point_size = 2,
  ens_n = 1,
  add_obs = TRUE,
  level = FALSE,
  remove_spin_up = TRUE,
  print_plots = FALSE,
  var_lims = NULL,
  ylim = NULL,
  cumulative = FALSE,
  facet = TRUE
)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; model(s) to plot. If missing, all models
in the Aeme object will be plotted.}

\item{var_sim}{string; of variable to plot}

\item{ens_n}{integer; ensemble number to plot. Default is 1.}

\item{add_obs}{logical; add observations to plot}

\item{level}{logical; include lake level. Only applies for contour plots.}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}

\item{print_plots}{logical; print plots}

\item{var_lims}{numeric vector of length 2; limits for the variable.
Defaults to NULL and will generate common limits for all variables.}

\item{ylim}{numeric vector of length 2; limits for the y-axis. Defaults to
NULL and calculates this based on the data to be plotted.}

\item{cumulative}{logical; plot cumulative sum of variable}

\item{facet}{logical; if \code{TRUE}, for variables with depth, plot each
model in a separate facet. If \code{FALSE}, plot each model in a separate
plot and return a list of plots. If \code{TRUE}, for variables without depth,
plot each model in a separate facet. If \code{FALSE}, plot each model as a
separate line and return a plot.
This only applies to variables without a depth component.}
}
\value{
list of plots for z-dimensional variables or a ggplot2 object for 1-d
variables.
}
\description{
Generates a contour plot for z-dimension variables and line plot for 1-d
variables.
}
\examples{
\dontrun{
  tmpdir <- tempdir()
  aeme_dir <- system.file("extdata/lake/", package = "AEME")
  # Copy files from package into tempdir
  file.copy(aeme_dir, tmpdir, recursive = TRUE)
  path <- file.path(tmpdir, "lake")
  aeme <- yaml_to_aeme(path = path, "aeme.yaml")
  model_controls <- get_model_controls()
  inf_factor = c("dy_cd" = 1, "glm_aed" = 1, "gotm_wet" = 1)
  outf_factor = c("dy_cd" = 1, "glm_aed" = 1, "gotm_wet" = 1)
  #'   model <- c("glm_aed", "gotm_wet")
  build_aeme(path = path, aeme = aeme, model = model,
                 model_controls = model_controls, inf_factor = inf_factor, ext_elev = 5,
                 use_bgc = TRUE)
  run_aeme(aeme = aeme, model = model, verbose = FALSE, path = path,
           parallel = TRUE)

  aeme <- load_output(model = model, aeme = aeme, path = path,
                           model_controls = model_controls, parallel = TRUE)

  p1 <- plot_output(aeme = aeme, model = model,
                    var_sim = "HYD_temp", level = TRUE,
                    print_plots = FALSE, var_lims = c(0, 30),
                    ylim = c(0, 16))
  p1[[1]]

  p2 <- plot_output(aeme = aeme, model = model,
                    var_sim = "LKE_evpvol", print_plots = TRUE,
                    ylim = c(0, 0.02))
}

}
