% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{plot_output}
\alias{plot_output}
\title{Plot AEME output}
\usage{
plot_output(
  aeme_data,
  model,
  var_sim = "HYD_temp",
  add_obs = TRUE,
  level = FALSE,
  label = FALSE,
  print_plots = TRUE,
  var_lims = NULL,
  ylim = NULL,
  cumulative = FALSE,
  facet = FALSE
)
}
\arguments{
\item{aeme_data}{aeme; data object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{var_sim}{string; of variable to plot}

\item{add_obs}{boolean; add observations to plot}

\item{level}{boolean; include lake level. Only applies for contour plots.}

\item{label}{boolean; include variable label as subtitle for contour plots.}

\item{print_plots}{boolean; print plots}

\item{var_lims}{numeric vector of length 2; limits for the variable.
Defaults to NULL and will generate common limits for all variables.}

\item{ylim}{numeric vector of length 2; limits for the y-axis. Defaults to
NULL and calculates this based on the data to be plotted.}
}
\value{
list of plots for z-dimensional variables or a ggplot2 object for 1-d
variables.
}
\description{
Generates a contour plot for z-dimension variables and line plot for 1-d
variables.
}
\examples{
\dontrun{
  tmpdir <- tempdir()
  aeme_dir <- system.file("extdata/lake/", package = "AEME")
  # Copy files from package into tempdir
  file.copy(aeme_dir, tmpdir, recursive = TRUE)
  path <- file.path(tmpdir, "lake")
  aeme_data <- yaml_to_aeme(path = path, "aeme.yaml")
  mod_ctrls <- read.csv(file.path(path, "model_controls.csv"))
  inf_factor = c("dy_cd" = 1, "glm_aed" = 1, "gotm_wet" = 1)
  outf_factor = c("dy_cd" = 1, "glm_aed" = 1, "gotm_wet" = 1)
  #'   model <- c("glm_aed", "gotm_wet")
  build_ensemble(path = path, aeme_data = aeme_data, model = model,
                 mod_ctrls = mod_ctrls, inf_factor = inf_factor, ext_elev = 5,
                 use_bgc = TRUE)
  run_aeme(aeme_data = aeme_data, model = model, verbose = FALSE, path = path,
           parallel = TRUE)

  aeme_data <- load_output(model = model, aeme_data = aeme_data, path = path,
                           mod_ctrls = mod_ctrls, parallel = TRUE)

  p1 <- plot_output(aeme_data = aeme_data, model = model,
                    var_sim = "HYD_temp", level = TRUE, label = TRUE,
                    print_plots = FALSE, var_lims = c(0, 30),
                    ylim = c(0, 16))
  p1[[1]]

  p2 <- plot_output(aeme_data = aeme_data, model = model,
                    var_sim = "LKE_evpvol", print_plots = TRUE,
                    ylim = c(0, 0.02))
}

}
