% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phs.R
\name{plot_phs}
\alias{plot_phs}
\title{Plot phytoplankton variables}
\usage{
plot_phs(aeme, model, add_obs = TRUE, depth_range = NULL, ens_n = 1)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{depth_range}{numeric; range of depths to plot. Default is NULL, which
averages over all depths.}
}
\value{
A ggplot object
}
\description{
Plot phytoplankton variables
}
