% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resid.R
\name{plot_resid}
\alias{plot_resid}
\title{Plot model residuals}
\usage{
plot_resid(aeme, model, var_sim = "HYD_temp")
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{character vector; model(s) to plot. If missing, all models
in the Aeme object will be plotted.}

\item{var_sim}{string; of variable to plot}
}
\value{
list of ggplot2 objects
}
\description{
Plot model residuals
}
