% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot multi-variable timeseries}
\usage{
plot_ts(aeme, model, var_sim, add_obs = TRUE, depth_range = NULL, ens_n = 1)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{var_sim}{character; variable in the AEME format (e.g. "HYD_temp"). Can
be a vector.}

\item{depth_range}{numeric; range of depths to average. Default is NULL,
which averages over all depths.}
}
\value{
A ggplot object
}
\description{
Plot multi-variable timeseries
}
