% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var.R
\name{plot_var}
\alias{plot_var}
\title{Plot AEME variable}
\usage{
plot_var(
  df = NULL,
  aeme,
  model,
  var_sim,
  ylim = NULL,
  xlim,
  var_lims = NULL,
  obs = NULL,
  add_obs = TRUE,
  point_size = 2,
  level = FALSE,
  facet = FALSE,
  cumulative = FALSE,
  print_plots = FALSE
)
}
\arguments{
\item{df}{dataframe; output from \code{\link{get_var}}. If \code{NULL},
\code{\link{get_var}} will be called and \code{var_sim} will be used to
extract the variable of interest from \code{aeme} for each model.}

\item{aeme}{Aeme object.}

\item{model}{character vector; model(s) to plot. If missing, all models
in the Aeme object will be plotted.}

\item{var_sim}{string; of variable to plot}

\item{ylim}{numeric vector of length 2; limits for the y-axis. Defaults to
NULL and calculates this based on the data to be plotted.}

\item{xlim}{numeric; x-axis limits}

\item{var_lims}{numeric vector of length 2; limits for the variable.
Defaults to NULL and will generate common limits for all variables.}

\item{obs}{list; output from \code{\link{observations}}}

\item{add_obs}{logical; add observations to plot}

\item{level}{logical; include lake level. Only applies for contour plots.}

\item{facet}{logical; if \code{TRUE}, for variables with depth, plot each
model in a separate facet. If \code{FALSE}, plot each model in a separate
plot and return a list of plots. If \code{TRUE}, for variables without depth,
plot each model in a separate facet. If \code{FALSE}, plot each model as a
separate line and return a plot.
This only applies to variables without a depth component.}

\item{cumulative}{logical; plot cumulative sum of variable}

\item{print_plots}{logical; print plots}
}
\value{
ggplot2 object or list of ggplot2 objects
}
\description{
Plot AEME variable
}
