% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbal_annual.R
\name{plot_wbal_annual}
\alias{plot_wbal_annual}
\title{Plot annual water balance components}
\usage{
plot_wbal_annual(aeme, remove_spin_up = FALSE)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
ggplot2 object
}
\description{
Calculates and plots annual water balance components including evaporation,
precipitation, inflow, outflow, and net balance for each model in the Aeme
 object.
}
