% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbal_comp.R
\name{plot_wbal_comp}
\alias{plot_wbal_comp}
\title{Plot Water Balance Comparison}
\usage{
plot_wbal_comp(wbal)
}
\arguments{
\item{wbal}{Water balance object created with \code{get_wbal_components()}}
}
\value{
ggplot object
}
\description{
Plot Water Balance Comparison
}
