% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbal_summaries.R
\name{plot_wbal_summaries}
\alias{plot_wbal_summaries}
\title{Plot Water Balance Summaries}
\usage{
plot_wbal_summaries(aeme)
}
\arguments{
\item{aeme}{Aeme object with model output and observations added. Must have
model output \code{run_aeme()}}
}
\value{
A ggplot2 object with water balance summaries
}
\description{
Plot Water Balance Summaries
}
