% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weir_calibration.R
\name{plot_weir_calibration}
\alias{plot_weir_calibration}
\title{Visualise calibrated weir parameters (C, h_inv) from calc_water_balance().}
\usage{
plot_weir_calibration(aeme, model)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}
}
\value{
A patchwork object.
}
\description{
Visualise calibrated weir parameters (C, h_inv) from calc_water_balance().
}
