% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weir_calibration.R
\name{plot_weir_calibration}
\alias{plot_weir_calibration}
\title{Visualise calibrated weir parameters (C, h_inv) from calc_water_balance().}
\usage{
plot_weir_calibration(aeme, model)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}, \code{"simstrat_aed2"}. Defaults to all models if
not found in \code{aeme}.}
}
\value{
A patchwork object.
}
\description{
Visualise calibrated weir parameters (C, h_inv) from calc_water_balance().
}
