% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wlev.R
\name{plot_wlev}
\alias{plot_wlev}
\title{Plot lake water level}
\usage{
plot_wlev(aeme, model, facet = FALSE, ...)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{character vector; model(s) to plot. If missing, all models
in the Aeme object will be plotted.}

\item{facet}{logical; if \code{TRUE}, for variables with depth, plot each
model in a separate facet. If \code{FALSE}, plot each model in a separate
plot and return a list of plots. If \code{TRUE}, for variables without depth,
plot each model in a separate facet. If \code{FALSE}, plot each model as a
separate line and return a plot.
This only applies to variables without a depth component.}

\item{...}{Additional arguments passed to \code{plot_output}}
}
\value{
A ggplot object
}
\description{
Plot lake water level
}
