% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zoops.R
\name{plot_zoops}
\alias{plot_zoops}
\title{Plot phytoplankton variables}
\usage{
plot_zoops(
  aeme,
  model,
  add_obs = TRUE,
  depth_range = NULL,
  remove_spin_up = TRUE,
  ens_n = 1
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{depth_range}{numeric; range of depths to plot. Default is NULL, which
averages over all depths.}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}

\item{ens_n}{numeric; ensemble number to allocate to model output which is
loaded. Defaults to 1.}
}
\value{
A ggplot object
}
\description{
Plot phytoplankton variables
}
