% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_status.R
\name{precip_status}
\alias{precip_status}
\title{Get current precipitation status in Aeme object}
\usage{
precip_status(aeme)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
character. Either "precip_as_met", "precip_as_inflow" or "no_precip"
}
\description{
This function checks whether precipitation is currently set as a meteorological input
or as an inflow in the Aeme object. It examines the meteorological data for
precipitation values and the inflow data for a precipitation inflow.
}
