% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aed_param_csv.R
\name{read_aed_param_csv}
\alias{read_aed_param_csv}
\alias{write_aed_param_csv}
\title{Read and write AED parameter CSV files
These functions handle reading and writing AED parameter CSV files,
ensuring that single quotes are properly managed in column names and
specific columns.}
\usage{
read_aed_param_csv(file)

write_aed_param_csv(df, file)
}
\arguments{
\item{file}{The path to the CSV file.}

\item{df}{A data frame to be written to CSV.}
}
\value{
For `read_aed_param_csv`, a data frame read from the CSV file.
}
\description{
Read and write AED parameter CSV files
These functions handle reading and writing AED parameter CSV files,
ensuring that single quotes are properly managed in column names and
specific columns.
}
