% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aeme_from_files.R
\name{read_aeme_from_files}
\alias{read_aeme_from_files}
\title{Read an AEME object from files}
\usage{
read_aeme_from_files(path)
}
\arguments{
\item{path}{Path to the directory containing the AEME files.}
}
\value{
An AEME object populated with data from the files.
}
\description{
Read an AEME object from files
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
path <- "test_write"
model_controls <- get_model_controls()
aeme <- build_aeme(path = path, aeme = aeme, model = "glm_aed",
model_controls = model_controls, ext_elev = 3)
aeme <- run_aeme(aeme = aeme, model = "glm_aed", path = path)
write_aeme_to_files(aeme, path)
aeme_path <- get_lake_dir(aeme = aeme, path = path)
aeme2 <- read_aeme_from_files(aeme_path)
}
