% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_glm_output.R
\name{read_glm_output}
\alias{read_glm_output}
\title{Read GLM netCDF output}
\usage{
read_glm_output(
  nc = NULL,
  vars_sim = NULL,
  depths = NULL,
  dates = NULL,
  date_index = NULL,
  incl_fluxes = TRUE,
  output_hour = 0,
  file,
  phyto_pars
)
}
\arguments{
\item{nc}{An object of class \code{ncdf4} (as returned by either 
 function \code{\link[ncdf4]{nc_open}}
 or function \code{\link[ncdf4]{nc_create}}), indicating what file to read from.}

\item{vars_sim}{Variables to extract in the AEME format e.g. "HYD_temp"}

\item{depths}{Depths to extract. If NULL, extract all model layer depths. 
Defaults to NULL.}

\item{dates}{Dates to extract. If NULL, extract all dates. Defaults to NULL.}

\item{date_index}{Date index to extract. If NULL, extract all dates. Defaults
to NULL.}

\item{incl_fluxes}{Logical indicating whether to include flux variables.
Defaults to TRUE.}

\item{output_hour}{Hour of the day to extract (0-23). Defaults to 0.}

\item{file}{File path to netCDF file. Only used if `nc` is NULL.}

\item{phyto_pars}{Data frame with phytoplankton parameters from AED.}
}
\value{
List with AEME output variables
}
\description{
Read GLM netCDF output
}
