% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_glm_output.R
\name{read_glm_wlev}
\alias{read_glm_wlev}
\title{Read GLM lake water level output}
\usage{
read_glm_wlev(nc = NULL, file)
}
\arguments{
\item{nc}{An object of class \code{ncdf4} (as returned by either 
 function \code{\link[ncdf4]{nc_open}}
 or function \code{\link[ncdf4]{nc_create}}), indicating what file to read from.}

\item{file}{File path to netCDF file. Only used if `nc` is NULL.}
}
\value{
Data frame with Date and LKE_lvlwtr columns
}
\description{
Read GLM lake water level output
}
