% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gotm_flux_output.R
\name{read_gotm_flux_output}
\alias{read_gotm_flux_output}
\title{Read GOTM flux output}
\usage{
read_gotm_flux_output(nc = NULL, file, dates = NULL, date_index = NULL)
}
\arguments{
\item{nc}{An object of class \code{ncdf4} (as returned by either 
 function \code{\link[ncdf4]{nc_open}}
 or function \code{\link[ncdf4]{nc_create}}), indicating what file to read from.}

\item{file}{File path to netCDF file. Only used if \code{nc} is NULL.}

\item{dates}{Dates to extract. If NULL, extract all dates. Defaults to NULL.}

\item{date_index}{Date index to extract. If NULL, extract all dates. Defaults
to NULL.}
}
\value{
List with GOTM flux output variables
}
\description{
Read GOTM flux output
}
