% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_wlev.R
\name{read_model_wlev}
\alias{read_model_wlev}
\title{Read water level from model output}
\usage{
read_model_wlev(nc = NULL, lake_dir, model)
}
\arguments{
\item{nc}{Open netCDF object. If NULL, will open netCDF from lake_dir. This
is useful when reading multiple variables from the same file to avoid
reopening the file multiple times. Defaults to NULL.}

\item{lake_dir}{Directory of lake model outputs}

\item{model}{Model name. One of "gotm_wet", "glm_aed", or "dy_cd".}
}
\value{
A data frame with columns:
\itemize{
\item \code{Date}: Date-time of the water level observation (POSIXct, UTC)
\item \code{LKE_lvlwtr}: Water level (meters)
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function reads water level data from the output of supported lake models.
}
