% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_simstrat_output.R
\name{read_simstrat_wlev}
\alias{read_simstrat_wlev}
\title{Read Simstrat-AED2 lake water level output}
\usage{
read_simstrat_wlev(nc = NULL, file)
}
\arguments{
\item{nc}{An object of class \code{ncdf4} (as returned by either 
 function \code{\link[ncdf4]{nc_open}}
 or function \code{\link[ncdf4]{nc_create}}), indicating what file to read from.}

\item{file}{File path to netCDF file. Only used if \code{nc} is NULL.}
}
\value{
Data frame with Date and LKE_lvlwtr columns
}
\description{
Read Simstrat-AED2 lake water level output
}
