% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_inflow.R
\name{remove_inflow}
\alias{remove_inflow}
\title{Remove inflow data from Aeme object}
\usage{
remove_inflow(aeme, inflow_id)
}
\arguments{
\item{aeme}{aeme; object.}

\item{inflow_id}{A string specifying the identifier for the inflow.}
}
\value{
Aeme object with inflow removed
}
\description{
Remove specified inflow data from an Aeme object.
}
