% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_output.R
\name{remove_output}
\alias{remove_output}
\title{Remove output from AEME object}
\usage{
remove_output(aeme)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
aeme object
}
\description{
Remove output from AEME object
}
