% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_param.R
\name{remove_param}
\alias{remove_param}
\title{Remove parameter(s) from Aeme object}
\usage{
remove_param(aeme, name)
}
\arguments{
\item{aeme}{Aeme object.}

\item{name}{character vector with names of parameters to remove. If missing,
all parameters are removed.}
}
\value{
Aeme object with parameters removed
}
\description{
Remove parameter(s) from Aeme object
}
