% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme.R
\name{run_aeme}
\alias{run_aeme}
\title{Run aquatic model ensemble}
\usage{
run_aeme(
  config,
  model,
  verbose = FALSE,
  debug = FALSE,
  timeout = 0,
  parallel = FALSE,
  dir = "."
)
}
\arguments{
\item{config}{list; loaded via `config <- configr::read.config("aeme.yaml")`}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{verbose}{boolean; print model output to console. Defaults to FALSE.}

\item{debug}{boolean; write debug log (Only DYRESM). Defaults to FALSE.}

\item{timeout}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}

\item{parallel}{boolean; run models in parallel. Defaults to FALSE.}

\item{dir}{filepath; where input files are located relative to `config`.}
}
\value{
Runs the model
}
\description{
Run aquatic model ensemble
}
