% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ensemble.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run ensemble}
\usage{
run_ensemble(model, lake_dir, bin_path, verbose = FALSE, parallel = FALSE)
}
\arguments{
\item{parallel}{}
}
\value{
Runs the model
}
\description{
Run ensemble
}
