% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aed_sed_const2d.R
\name{set_aed_sed_const2d}
\alias{set_aed_sed_const2d}
\title{Setup AED aed_sed_const2d sediment model parameters}
\usage{
set_aed_sed_const2d(
  aeme,
  path,
  lake_dir = NULL,
  baseline = c(fsed_oxy = -25, fsed_amm = 2, fsed_nit = 0.2, fsed_frp = 0.05)
)
}
\arguments{
\item{aeme}{Aeme object.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}

\item{baseline}{Named numeric vector of baseline fluxes at
\code{ref_depth}. Must include \code{fsed_oxy}, \code{fsed_amm},
\code{fsed_nit}, \code{fsed_frp}.}
}
\value{
Invisible NULL. Updates the aed.nml file in the glm_aed model
directory.
}
\description{
Setup AED aed_sed_const2d sediment model parameters
}
