% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aed_sed_const2d.R
\name{set_aed_sed_const2d}
\alias{set_aed_sed_const2d}
\title{Setup AED aed_sed_const2d sediment model parameters}
\usage{
set_aed_sed_const2d(aeme, path, lake_dir = NULL)
}
\arguments{
\item{aeme}{aeme; object.}

\item{path}{filepath; where input files are located relative to the current 
working directory.}
}
\value{
Invisible NULL. Updates the aed.nml file in the glm_aed model
 directory.
}
\description{
Setup AED aed_sed_const2d sediment model parameters
}
