% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_inflows.R
\name{set_inflows}
\alias{set_inflows}
\title{Set inflow data from Aeme object}
\usage{
set_inflows(aeme, inflows)
}
\arguments{
\item{aeme}{aeme; object.}

\item{inflows}{Either a named list of data frames for each inflow, or a
single data frame containing an "inflow_id" column to identify different
inflows. Each data frame must have columns "Date", "HYD_flow", "HYD_temp" and
"CHM_salt".}
}
\value{
Aeme object with inflows replaced
}
\description{
Replace existing inflow data in an Aeme object with new inflow data. The new
inflow data can be provided as either a named list of data frames or a single
data frame containing an "inflow_id" column to identify different inflows.
}
