% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_inflows.R
\name{set_inflows}
\alias{set_inflows}
\title{Set inflow data from Aeme object}
\usage{
set_inflows(aeme, inflows)
}
\arguments{
\item{aeme}{aeme; object.}
}
\value{
Aeme object with inflows replaced
}
\description{
Replace existing inflow data in an Aeme object with new inflow data. The new
inflow data can be provided as either a named list of data frames or a single
data frame containing an "inflow_id" column to identify different inflows.
}
