% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_model_controls.R
\name{set_model_controls}
\alias{set_model_controls}
\title{Set model controls for AEME object}
\usage{
set_model_controls(aeme, model_controls)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}
}
\value{
Aeme object with model controls set
}
\description{
Set model controls for AEME object
}
