% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_precip.R
\name{set_precip}
\alias{set_precip}
\title{Set precipitation as either meteorological input or inflow}
\usage{
set_precip(
  aeme,
  type = c("inflow", "met", "precip_as_inflow", "precip_as_met")
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{type}{character. Either "met" or "inflow". Default is
"inflow". If "met", precipitation is treated as a meteorological input
in mm. If "inflow", precipitation is converted to an inflow volume in m3. 
It is also possible to use the old argument values "precip_as_met" and 
"precip_as_inflow" for backward compatibility.}
}
\value{
Aeme object with precipitation set as specified
}
\description{
This function allows you to specify whether precipitation should be treated as 
a meteorological input (in mm) or as an inflow volume (in m3) to the lake.
When set as an inflow, the function calculates the inflow volume based on the
lake surface area and the precipitation amount. Conversely, when set as a
meteorological input, it converts the inflow volume back to precipitation depth.
This is useful for hydrodynamic modeling where the representation of precipitation
can impact the simulation results, particularly in terms of water balance and
thermal structure.
}
