% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_time.R
\name{set_time}
\alias{set_time}
\title{Set time parameters for an Aeme object}
\usage{
set_time(aeme, start, stop, spin_up)
}
\arguments{
\item{aeme}{aeme; object.}

\item{start, stop}{Time in the format "YYYY-mm-dd" or "YYYY-mm-dd HH:MM" or
"YYYY-mm-dd HH:MM:SS"}

\item{spin_up}{Spin-up time in days. Can be a single numeric value or a list with
model names as names and numeric values as values.}
}
\value{
Aeme object with time parameters set
}
\description{
Set time parameters for an Aeme object
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
aeme <- set_time(aeme = aeme, start = "2020-01-01", stop = "2020-12-31",
                 spin_up = 35)
}
