% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vars_sim.R
\name{set_vars_sim}
\alias{set_vars_sim}
\title{Set simulation variables in model controls}
\usage{
set_vars_sim(model_controls, vars_sim, simulate = TRUE, exclusive = FALSE)
}
\arguments{
\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}

\item{vars_sim}{character vector of variable names to set for simulation}

\item{simulate}{logical, whether to simulate the variables in vars_sim}

\item{exclusive}{logical, if TRUE, set all other variables to not simulate}
}
\value{
Updated model_controls data frame
}
\description{
Set simulation variables in model controls
}
\examples{
model_controls <- data.frame(
 var_aeme = c("HYD_temp", "CHM_oxy", "PHS_tp"),
 simulate = c(TRUE, FALSE, TRUE),
 inf_default = c(NA, NA, NA),
 initial_wc = c(NA, NA, NA),
 initial_sed = c(NA, NA, NA),
 conversion_aed = c(1, 1, 1),
 stringsAsFactors = FALSE
)
vars_sim <- c("CHM_oxy", "NIT_tn", "PHS_tp")
updated_controls <- set_vars_sim(model_controls, vars_sim, simulate = TRUE)
print(updated_controls)

}
