% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{show,aeme-method}
\alias{show,aeme-method}
\title{Print aeme object to the console}
\usage{
\S4method{show}{aeme}(object)
}
\arguments{
\item{object}{An aeme object.}
}
\value{
prints the aeme object to the console.
}
\description{
Show an aeme object in the console
}
\details{
This method prints the aeme output in a readable format to the console.
}
