% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{summary,Aeme-method}
\alias{summary,Aeme-method}
\title{Summarise an Aeme object}
\usage{
\S4method{summary}{Aeme}(object)
}
\arguments{
\item{object}{An Aeme object.}
}
\value{
Aeme object.
}
\description{
Summarise an Aeme object
}
\details{
This method summarises the Aeme output.
}
