% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{time,aeme-method}
\alias{time,aeme-method}
\title{Access time slot}
\usage{
\S4method{time}{aeme}(aeme)
}
\arguments{
\item{aeme}{A aeme object.}
}
\value{
list of time characteristics
}
\description{
Access time slot
}
