% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{time<-,Aeme-method}
\alias{time<-,Aeme-method}
\title{Set time in Aeme object}
\usage{
\S4method{time}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New time data to be assigned.}
}
\value{
A modified Aeme object with updated time slot.
}
\description{
Update the time slot of an Aeme object
}
\details{
This method updates the "time" slot of An Aeme object with new data.
}
