% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{time<-}
\alias{time<-}
\title{Set time in aeme object}
\usage{
time(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New time data to be assigned.}
}
\value{
A modified aeme object with updated time slot.
}
\description{
Update the time slot of an aeme object
}
