% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{time<-}
\alias{time<-}
\title{Set time in Aeme object}
\usage{
time(aeme) <- value
}
\arguments{
\item{aeme}{A Aeme object.}

\item{value}{New time data to be assigned.}
}
\value{
A modified Aeme object with updated time slot.
}
\description{
Update the time slot of an Aeme object
}
