% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{time}
\alias{time}
\alias{time,Aeme-method}
\title{Access the time slot of an Aeme object}
\usage{
time(aeme)

\S4method{time}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
List of time characteristics.
}
\description{
Access the time slot of an Aeme object
}
