% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_models.R
\name{toggle_models}
\alias{toggle_models}
\title{Toggle between model display names and codes}
\usage{
toggle_models(model = NULL, to = NULL)
}
\arguments{
\item{model}{Model name to be toggled.}

\item{to}{Target format: either "display" for user-friendly names or "code"
for backend codes. If NULL, the function will toggle to the opposite format.}
}
\value{
The model name in the desired format.
}
\description{
A helper function to switch between user-friendly model names and their
corresponding codes used in the backend.
}
\examples{
toggle_models("DYRESM-CAEDYM")  # Returns "dy_cd"
toggle_models("dy_cd", to = "display")  # Returns "DYRESM-CAEDYM"
toggle_models("GLM-AED")  # Returns "glm_aed"
toggle_models("gotm_wet", to = "display")  # Returns "GOTM-WET"
toggle_models("GOTM-WET", to = "code")  # Returns "gotm_wet"
df <- data.frame(model = rep(c("dy_cd", "glm_aed"), 50))
df <- df |>
        dplyr::mutate(Model = toggle_models(model, to = "display"))
}
