% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_aeme_vars.R
\name{view_aeme_vars}
\alias{view_aeme_vars}
\title{View AEME variables}
\usage{
view_aeme_vars(group = "all")
}
\arguments{
\item{group}{character string to filter variables by group. Default is "all".
Options include "HYD" (hydrodynamic), "LKE" (Lake), "RAD" (Radiation),
"CHM" (Chemistry), "PHS" (Phosphorus), "NIT" (Nitrogen), "CAR" (Carbon),
 "SIL" (Silica), "MET" (Meteorologica), "ZOO" (Zooplankton),
 "PHY" (Phytoplankton),}
}
\value{
A data frame with AEME variables
}
\description{
View AEME variables
}
\examples{
view_aeme_vars()
view_aeme_vars("HYD")
view_aeme_vars("CHM")
view_aeme_vars("PHY")
}
