% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{water_balance,Aeme-method}
\alias{water_balance,Aeme-method}
\title{Access water_balance slot}
\usage{
\S4method{water_balance}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
list of water_balance characteristics
}
\description{
Access water_balance slot
}
