% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{water_balance<-,Aeme-method}
\alias{water_balance<-,Aeme-method}
\title{Set water_balance in Aeme object}
\usage{
\S4method{water_balance}{Aeme}(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New water_balance data to be assigned.}
}
\value{
A modified Aeme object with updated water_balance slot.
}
\description{
Update the water_balance slot of an Aeme object
}
\details{
This method updates the "water_balance" slot of a Aeme object with new data.
}
