% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{water_balance<-}
\alias{water_balance<-}
\title{Set water_balance in Aeme object}
\usage{
water_balance(aeme) <- value
}
\arguments{
\item{aeme}{An Aeme object.}

\item{value}{New water_balance data to be assigned.}
}
\value{
A modified Aeme object with updated water_balance slot.
}
\description{
Update the water_balance slot of an Aeme object
}
