% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{water_balance<-}
\alias{water_balance<-}
\title{Set water_balance in aeme object}
\usage{
water_balance(aeme) <- value
}
\arguments{
\item{aeme}{A aeme object.}

\item{value}{New water_balance data to be assigned.}
}
\value{
A modified aeme object with updated water_balance slot.
}
\description{
Update the water_balance slot of an aeme object
}
