% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeme.R
\name{water_balance}
\alias{water_balance}
\alias{water_balance,Aeme-method}
\title{Access the water_balance slot of an Aeme object}
\usage{
water_balance(aeme)

\S4method{water_balance}{Aeme}(aeme)
}
\arguments{
\item{aeme}{An Aeme object.}
}
\value{
List of water_balance characteristics.
}
\description{
Access the water_balance slot of an Aeme object
}
