% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aeme.R
\name{water_balance}
\alias{water_balance}
\title{Access water_balance slot}
\usage{
water_balance(aeme)
}
\arguments{
\item{aeme}{A Aeme object.}
}
\value{
list of water_balance characteristics
}
\description{
Access water_balance slot
}
