% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aeme_to_files.R
\name{write_aeme_to_files}
\alias{write_aeme_to_files}
\title{Write an AEME object to files}
\usage{
write_aeme_to_files(aeme, path, include_output = FALSE)
}
\arguments{
\item{aeme}{aeme; object.}

\item{path}{filepath; where input files are located relative to the current 
working directory.}

\item{include_output}{logical, include output files. Default is FALSE. The
output files can be large and take up a lot of space.}
}
\value{
A vector of file paths to the written files
}
\description{
Write an AEME object to files
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
model <- "glm_aed"
path <- "test_write"
model_controls <- get_model_controls()
aeme <- build_aeme(path = path, aeme = aeme, model = model,
model_controls = model_controls, ext_elev = 5)
aeme <- run_aeme(aeme = aeme, model = "glm_aed", path = path)
write_aeme_to_files(aeme, path)
}
