% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aeme_to_files.R
\name{write_aeme_to_files}
\alias{write_aeme_to_files}
\title{Write an AEME object to files}
\usage{
write_aeme_to_files(aeme, path)
}
\arguments{
\item{aeme}{aeme; object.}

\item{path}{filepath; where input files are located relative to `config`.}
}
\description{
Write an AEME object to files
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
path <- "test_write"
model_controls <- get_model_controls()
aeme <- build_aeme(path = path, aeme = aeme, model = "glm_aed",
model_controls = model_controls)
aeme <- run_aeme(aeme = aeme, model = "glm_aed", path = path)
write_aeme_to_files(aeme, path)
}
