% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_configuration.R
\name{write_configuration}
\alias{write_configuration}
\title{Write model configuration from the aeme object}
\usage{
write_configuration(aeme, model, path)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be \code{dy_cd}, \code{glm_aed},
\code{gotm_wet}.}

\item{path}{filepath; where input files are located relative to the current
working directory.}
}
\value{
aeme object which was passed to the function,
}
\description{
Write model configuration from the aeme object
}
