% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_configuration.R
\name{write_configuration}
\alias{write_configuration}
\title{Write model configuration from the aeme object}
\usage{
write_configuration(aeme, model, path)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}
}
\value{
aeme object which was passed to the function,
}
\description{
Write model configuration from the aeme object
}
