% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml.R
\name{write_yaml}
\alias{write_yaml}
\title{Write a yaml object to file}
\usage{
write_yaml(x, file, fileEncoding = "UTF-8", ...)
}
\arguments{
\item{x}{the object to be converted}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing}

\item{fileEncoding}{character string: if non-empty declares the
    encoding to be used on a file (not a connection) so the character data can
    be re-encoded as they are written.  See \code{\link{file}}.}

\item{...}{arguments to \code{\link[yaml]{as.yaml}}}
}
\description{
Write the YAML representation of an R object (list) to a file.
 Taken from the `yaml` package but added catch to replace 'yes' with 'true'
  and 'no' with 'false' and '~' with 'null' to make it compatible with
  running the GOTM model.
}
\examples{

\dontrun{
config_file <- system.file("extdata/feeagh/LakeEnsemblR.yaml", package = "LakeEnsemblR")

yaml <- read_yaml(config_file)
yaml <- set_yaml(yaml, value = 23, key1 = "location", key2 = "latitude")
yaml <- set_yaml(yaml, value = "2010-06-01 00:00:00", key1 = "time", key2 = "start")
yaml <- set_yaml(yaml, value = "meteo.csv", key1 = "input", key2 = "meteo", key3 = "file")
yaml <- set_yaml(yaml, value = TRUE, key1 = "calibration", key2 = "GOTM", key3 = "turb_param/k_min", key4 = "log")
yaml <- set_yaml(yaml, value = c("temp", "salt"), key1 = "output", key2 = "variables")

write_yaml(yaml, "LakeEnsemblR.yaml")
}
}
\author{
Jeremy Stephens <jeremy.f.stephens@vumc.org>, Tadhg Moore
}
