% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_aeme.R
\name{yaml_to_aeme}
\alias{yaml_to_aeme}
\title{Convert aeme.yaml file to list}
\usage{
yaml_to_aeme(path, file)
}
\arguments{
\item{path}{directory where aeme.yaml file is located. Can be used instead of \code{file} argument.}

\item{file}{filepath; to aeme.yaml file. Can be used instead of \code{path} and \code{file} arguments.}
}
\value{
aeme object
}
\description{
Convert aeme.yaml file to list
}
\examples{
aeme_yaml <- system.file("extdata/lake/aeme.yaml", package = "AEME")
aeme <- yaml_to_aeme(file = aeme_yaml)
aeme
}
