% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_aeme.R
\name{yaml_to_aeme}
\alias{yaml_to_aeme}
\title{Convert aeme.yaml file to list}
\usage{
yaml_to_aeme(path, file)
}
\arguments{
\item{path}{directory where aeme.yaml file is located. Can be used instead of \code{file} argument.}

\item{file}{filepath; to aeme.yaml file. Can be used instead of \code{path} and \code{file} arguments.}
}
\value{
aeme object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{yaml_to_aeme()} is soft-deprecated in favour of \code{\link[=aeme_constructor]{aeme_constructor()}} (build
an \code{Aeme} object from your own lake data, with full validation) or
\code{\link[=new_aeme]{new_aeme()}} (a quick placeholder object to populate incrementally). It
still works and will keep working, but new code should prefer those
instead of hand-editing a YAML file.
}
\examples{
aeme_yaml <- system.file("extdata/lake/aeme.yaml", package = "AEME")
aeme <- yaml_to_aeme(file = aeme_yaml)
aeme
}
\seealso{
\code{\link[=aeme_constructor]{aeme_constructor()}}, \code{\link[=new_aeme]{new_aeme()}}
}
