% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_aeme.R
\name{yaml_to_aeme}
\alias{yaml_to_aeme}
\title{Convert aeme.yaml file to list}
\usage{
yaml_to_aeme(dir, yaml_file)
}
\arguments{
\item{dir}{filepath; where input files are located relative to `config`.}

\item{yaml_file}{filepath; to aeme.yaml file}
}
\value{
aeme object
}
\description{
Convert aeme.yaml file to list
}
